Attempt to fix some outstanding bugs/misfeatures/bogons in the CVSROOT code, as well as simplifying some admin procedures.
The main change is to stop multi-directory commits only having the subject
info of the last directory, and not being sent to all the lists, and not
being logged to the appropriate commitlogs.

updateCVSROOT is gone (yay!), it's functionality is replaced by a cvs-1.6
feature - the 'checkoutlist' file, which is designed explicitly for this.
This will stop the easy-import users from getting told the mailing list
is out of date when a committer has been added but had missed out on being
put on the cvs-committers mailing list.

Update the comments and info in the *info files with the cvs-1.6 headers,
this fixes a few typos, clarifys some things, corrects some misleading
comments, etc.

Some stub files added to indicate how new cvs-1.6+ functionality is used,
but left commented out (taginfo, cvswrappers, etc)

All hard-coded references to /home/ncvs are changed back to $CVSROOT
(as suggested to me by bde) now that cvs-1.6+ makes a special effort
to ensure $CVSROOT is always set.  This allows the $CVSROOT files to be
used elsewhere (eg: for testing) with much less pain.

mkavail is history as well, it was often being forgotten.  cvs-committers
now takes the user list from /etc/group:ncvs via a trivial perl script.
This is so that we still have the ability to temporarily disable commit
access via 'avail' without cutting off the cvs-committers mailing list.
Since /etc/group editing is still required, this is not a new admin
overhead.  If we ever change so that membership of "ncvs" is not required
(eg: /usr/bin/cvs becomes setgid and access control is via avail) then
we are going to have to do it some other way.  (I have ideas)

I have tested this on my machines, but it is possible that I have missed
something.  As usual, please yell if I've botched it.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@2690 aed309b6-a8cd-e111-996c-001c23d10e55