Thanks for the overwhelming response (which can be only summarized by the word: "zilch").  I guess the only way to get people try and comment on these kind of things is to shove it down their throat.... ;)
Anyway, here's a set of changes required for auto-generation of READMEs
in ports directories.  Necessary changes and additions of templates
to the ports tree will follow shortly.

Eventually I'll commit all the generated READMEs to the tree, but that
will be in the rather distant future.  For now, I encourage anyone
with a -current systam and a matching ports tree to do a "make readmes"
at the top level and see what they get.

Next step will be to add pkg/{COMMENT,DESCR} to all the categories.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@2931 aed309b6-a8cd-e111-996c-001c23d10e55