A fancy-looking editor that's smaller than emacs (surprise!).
Note this port requires Motif, and is patched to build with libXm
static (to change this, just go hack in patches/patch-aa, it's only
two lines).  Some time later, I will make some changes to the ports
framework to make this configurable, but for now, it's "the way thud
wants it". :>

Submitted by:	"Daniel M. Eischen" <deischen@iworks.InterWorks.org>


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@2963 aed309b6-a8cd-e111-996c-001c23d10e55