Update to rdist-6.1.3.  The only major thing changed seems to be two buffer overflow problems, which shouldn't be an issue since it's not setuid.
While here, do not hardcode /usr/local, use $(PREFIX).  Do not patch
the install target so that 'make install' does the man pages, add
install.man to the ${INSTALL_TARGET} variable so that both "install"
and "install.man" are done. This reduces the diff to the base source a
little, but has the same overall effect.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@3828 aed309b6-a8cd-e111-996c-001c23d10e55