Suggested by: msmith Add a little more sophistication to the md5 grep command.
Change the md5 checksum logic a bit.  Now, the message is printed
out for every successfully/unsuccessfully matched checksum, and it
aborts at the end if there was a mismatch.  Also, make missing checksum
and IGNORE file inconsistency fatal, as there is now no reason to have
a missing checksum.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@5064 aed309b6-a8cd-e111-996c-001c23d10e55