Upgrade to 2.0.  More OS's, and now machiens are networked.
In playing this game, I found that our friend Chuck didn't come up as often
as before.  There are more OS's now, and Linux was represented by both
the Pinguin and Red Hat.  So I replaced "Red Hat" with a second "BSD".

The game often produces an error screen when bill wins.  But scores
are saved ok, and it plays fine.

Upgraded with permission of the Maintainer.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@5581 aed309b6-a8cd-e111-996c-001c23d10e55