Prevent this server error message:
    fatal: Local: Agent socket bind failed: Address already in use

It would happen when the server tried to create the Unix domain
socket "/tmp/ssh-username/agent-socket-123", if the file already
existed.  It could already exist if it happened to be left over from
a system crash.  This patch unlinks the file before attempting the
bind operation.

I will send this patch to ssh-bugs@cs.hut.fi too.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@7375 aed309b6-a8cd-e111-996c-001c23d10e55