Import of nttcp
This program is an up to date version of the ttcp program.
It uses inetd (or simulates its behaviour) to start off the remote
side program which will send/receive data. Both sides measure the time
and number of bytes transfered. The local side will print the measures.
The format of the output can be specified on the commandline.

PR:		ports/8546


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@15185 aed309b6-a8cd-e111-996c-001c23d10e55