Import of abacus sentry v 0.61
Sentry is part of the Abacus Project suite of security tools.
It is a program designed to detect and respond to port scans
against a target host in real-time. There are other port scan
detectors that peform similar detection of scans,but the Sentry
has some unique features that may make it worth looking into.

PR:		ports/5475
Submitted by:	chris@still.whet.org


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@15433 aed309b6-a8cd-e111-996c-001c23d10e55