bchunk is a Unix/C rewrite of the fine BinChunker software for some non-Unix systems.
binchunker converts a CD image in a ".bin / .cue" format
(sometimes ".raw / .cue") to a set of .iso and .cdr tracks. The
bin/cue format is used by some non-Unix cd-writing software, but
is not supported on most other cd-writing programs.

The .iso track contains an ISO file system, which can be mounted
through a vn device on FreeBSD systems, or written on a CD-R
using cdrecord. The .cdr tracks are in the native CD audio format. (pcm)
They can be either written on a CD-R using cdrecord -audio, or
converted to WAV (or any other sound format for that matter)
using sox.
PR:		ports/9420
Submitted By:	Adam McDougall <bsdx@spawnet.com>


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@16036 aed309b6-a8cd-e111-996c-001c23d10e55