Fix to make ports work with bsd.port.mk rev. 1.306. ${MACHINE_ARCH}--freebsd${OSREL} is now passed to CONFIGURE_ARGS if GNU_CONFIGURE is defined.  Take the target out of CONFIGURE_ARGS of some ports that added it explicitly; define it as ${MACHINE_ARCH}--freebsd if the port doesn't like the ${OSREL} part; define it as something else (such as ${MACHINE_ARCH}--freebsdelf if the port requires that; define it as an empty string if the port doesn't like it at all.
The last might be a sign that a GNU_CONFIGURE port actually doesn't
use GNU's version of configure at all; but I don't have time to go
look at them all, we'll fix them as time goes on.

At least we've got much fewer "-unknown-"s in the tree as the result. :)


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@17053 aed309b6-a8cd-e111-996c-001c23d10e55