Import of MM, a portable shared memory library.
The MM library is a 2-layer abstraction library which simplifies the usage
of shared memory between forked (and this way strongly related) processes
under Unix platforms. On the first layer it hides all platform dependent
implementation details (allocation and locking) when dealing with shared
memory segments and on the second layer it provides a high-level
malloc(3)-style API for a convenient and well known way to work with
data-structures inside those shared memory segments.

This library is proposed to be used in future versions
of Apache 1.3 as the base library for shared memory pools.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@17371 aed309b6-a8cd-e111-996c-001c23d10e55