Initial import of mod_perl 1.21.
This is mod_perl for Apache.  It is intended as a replacement for the
p5-Apache port.  Instead of creating a replacement httpd binary with
mod_perl linked into it, it uses Apache's APXS to create and install
mod_perl as a DSO.

This makes it suitable for use with apachexx-xxx ports other than
apache13.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@22923 aed309b6-a8cd-e111-996c-001c23d10e55