Import of corewars.
Core Wars is a simulation game in which a number of programs are loaded
into a virtual machine and executed.  The programs can try to crash each
other, manipulate each other, overwrite as much memory as they can, etc..
The "best" program is selected according to a scoring mechanism which
involves the number of memory cells the program has overwritten, the
number of other programs it has crashed, and whether/when it crashed.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@26460 aed309b6-a8cd-e111-996c-001c23d10e55