Initial import of lesstif-0.90.5 (linux version).
This port only installs library for linux-netscape(netscape-dynMotif).
Motif has a critical bug around multi-byte handling in TextWidget such as
<input type=text> and <textarea> fields.
This port will solve the problem (like www/libxpg4-ns for our native netscape).
Instead, it brings various problems (behaviour of scroll bar, a size of textarea, ..)
because lesstif is still in progress and not perfect clone of Motif.
If you have discontent with input/display words in linux-netscape, this is
worth trying this port.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@28327 aed309b6-a8cd-e111-996c-001c23d10e55