Do not try to use XML support since: 	(a)  ImageMagick needs libxml2, not libxml 	(b)  libxml2 does not co-exist AT ALL with libxml 	(c)  libxml is a required component of GNOME  	(d)  the extra functionality for ImageMagick is minimal
Thus, if you installed GNOME, and then this port, or vice-versa
Really Bad[tm] things will happen.

As and when libxml2 is used by GNOME (and libxml is buried
with partial military honors), then it will be possible to
turn this support back on.

Discovered by:	my crashbox going nuts after an install of GNOME
		and then ImageMagick.  There were no survivors :)


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@31236 aed309b6-a8cd-e111-996c-001c23d10e55