A small script to copy distfiles from per-port directory to central location.
This is done to avoid write-on-write race conditions, when two package
builds try to write the same distfile back into the server.  To avoid
this, the distfiles are first copied over to
distfiles/.pbtmp/${PKGNAME}, and then a .done file is created in that

directory.  This script runs on the server periodically to copy the
files to the central location.  Since this script copies one distfile
at a time, the race mentioned above doesn't exist.  Also, it will only
copy files from directories where the .done file exists to avoid
read-before-write race conditions.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@32075 aed309b6-a8cd-e111-996c-001c23d10e55