Finish updating to KDE2 snapshot 20000829A.  Re-add kdenetwork and kdeutils dependencies to kde2, and add koffice to kde2, because it's part of the 2.0 release.  Fix my bogon in forgetting to remove obsolete qt22 patches.
The current state of the ports as far as building and installing is quite
good, but I expect minor problems with PLISTs other than for the ones I've
fixed here.  I'm going to let bento figure it out.  :-)

Please, test these ports and report bugs to submit@bugs.kde.org, and send
me problems with the ports themselves.  I've found that KDE2 seems to have
stabilized significantly since the last snapshot.

And now, I am removing the last patch remaining in the KDE2 ports, except
for the QT 2.2 snapshot.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@32117 aed309b6-a8cd-e111-996c-001c23d10e55