Kludge rearming the signal for SIGCHLD so that we can harvest the zombies on the system.  Without this change, the socks5 daemon would work great for a short period of time, then hang because no more children could be created.  It appears from debugging that this is because gravedigger was only called once.  I reestablish the signal handler in gravedigger to force SIGCHLD to be unmasked.
A more proper fix might be to fix Signal to have multi-shot signals,
or to see what might be broken with the signal masks.  Since I'm in a
hurry and this port's daemon is broken without this, I commit this
kludge knowing that it will live approximately forever, less three
days.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@33348 aed309b6-a8cd-e111-996c-001c23d10e55