Gently persuade apache13-modssl to install in the same filesystem space as www/apache13 does, so that other apache-dependendent ports DTRT regardless of which one is installed.
Nothing concrete (other than we're working on a real cleanup of the
apache ports) has been heard from any of the maintainers.  Obviously
that is the best solution, but with 4.2-RELEASE just around the
corner, we have an opportunity here to at least remove a few bogons.

Both times I've raised this issue on -ports, however, there has been
significant interest in the patches, and now plenty of confirmations
that the modifications do not break the port in other ways.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@35170 aed309b6-a8cd-e111-996c-001c23d10e55