New port: enscript
This port adds Chinese support to enscript, which means it can convert
a BIG5/CNS plain text to a PS document, with all the fancy features
provided by enscript.  Unlike bg5ps, it can't manage GB yet, but it's
easy to add.  Another useful feature is with the ps2pdf and zh-ghostscript6,
it is now possible to generate font-embedded Chinese pdf document from
plain text on the fly, which also means platform-independent.

Due to some problems in the CMap files provided by Adobe, Eten-B5-H and
B5-H CMaps are unable to map correct ascii glyphs in TTFs.  BIG5 users
are suggested to use B5pc-H for the time being.

Maybe it's time to re-organize ports/chinese/* font relationships.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@35470 aed309b6-a8cd-e111-996c-001c23d10e55