Per the instructions of the software, cfengine is now dependent on BerkleyDB 3. In fact, the only thing (that I can tell anyway) that uses the DB functions is cfd. I'm guessing most people don't use it, especially since I haven't heard any complaints about it.
Also, it was putting some of the docs in / (for some really strange reason).
That should be fixed too. They are now put in {PREFIX}/share/doc/cfengine,
depending on NOPORTSDOC (somewhat reverse logic than normal, but it made the
amount of changes less).

Finally, strip the binaries.

PR:		26189
Submitted by:	    maintainer


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@40737 aed309b6-a8cd-e111-996c-001c23d10e55