Update modules: net and apache.
From the webmin updates page:

apache: Two virtual servers can be created with the same name and port. On
systems that restrict which Webmin users can edit which virtual servers,
this could be used to get around the restriction.

net: Virtual network interfaces could not be setup properly on Solaris 8.
(this should not affect FreeBSD users).


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@40846 aed309b6-a8cd-e111-996c-001c23d10e55