Usually when you have to store persistent data you don't need a full-blown database server, just a ASCII database would do the trick.
AsciiDB::Tag allows you to access a simple ASCII database using a
perl hash variable. The database format is straightforward so you
can edit it by hand if you need so. Each record is stored into a
file, and a record is just a set of values tagged by the field
name.

PR:	ports/26857


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@42003 aed309b6-a8cd-e111-996c-001c23d10e55