Initial import of wmlpq-0.2.1.
wmlpq is a Windowmaker dockapp which monitors up to 5 printqueues,
allowing you to see when your print job has finally been printed.
This is useful when you work in a networked environment with several
printers available, which are not in the same room; or (in some
cases) the same country, and is easier than typing lpq every 30
seconds.

PR: 28750
Submitted by: Patrick Li <pat@databits.net>


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@44935 aed309b6-a8cd-e111-996c-001c23d10e55