Initial import of twoftpd-1.13.
Twoftpd is an FTP server that strives to be secure, simple, and efficient.
None of the commands can cause execution of other programs, and the normal
model of execution does a chroot to the logged in user's directory
immediately after authentication.

PR: 30565
Submitted by: Gea-Suan Lin <gslin@infomath.math.nctu.edu.tw>


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@47872 aed309b6-a8cd-e111-996c-001c23d10e55