Update to version 0.7.
Version 0.6 has been skipped, as the author released 0.7 as integration of
the patches that would have been needed in the 0.6 port.

New feature: Support for limiting the number of messages downloaded in a
mailbox.  Ideal for reading the FreeBSD lists offline while keeping a
complete mirror on your IMAP server.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@50564 aed309b6-a8cd-e111-996c-001c23d10e55