Update to exim-4.01:
Quite a few bugfixes and two new features.  I'm not going to try to
cover changes in commit logs for this port; the software's still
in the early stages of its release cycle, and version-chasers should
be reading the ChangeLog and NewStuff documentation.

I've also changed the way the distribution version numbers are stored
so that I don't have to modify the pkg-plist every time I update the
port.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@55584 aed309b6-a8cd-e111-996c-001c23d10e55