Update to version 20020203.
Since the author switched to use autoconf to maintain it, the entire
build system changed quite a bit.  In order to get it to use the
correct tools (avr-gcc etc.) instead of the default C compiler, the
package now contains wrapper scripts to configure and build it, so a
good part of our ports framework doesn't apply anymore very well.  As
a side-effect, i can't get it to honor ${PREFIX} in any way right now.

Maybe i'll switch away from the shipped scripts, and redo all their
stuff within that Makefile in future.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@57864 aed309b6-a8cd-e111-996c-001c23d10e55