New port jboss2.
The jboss2ctl program does not work at this stage. I chose
to commit this port anyway because it has been lying around
for too long. Once it is in the ports tree, the improved
interest may trigger somebody else to fix the problem.

JBoss can be started using the supplied shell script anyway,
so it is still useful.

This is not the JBoss port submitted by Jesse Gallup in PR
35853, but one based on the www/orion port.

PR:	35853


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@58192 aed309b6-a8cd-e111-996c-001c23d10e55