Move back to share/gnome as a datadir.
The GNOME 2.0 APIs can be installed right along with the GNOME 1.4 APIs, but
the desktops will be mutually independent.  That is, you will not be
able to install gnomedesktop, for example, on the same machine as gnomecore.
However, you will be able to install the GNOME 2.0 desktop, and use all your
favorite GNOME 1.4 applications (like gaim, AbiWord, Evolution, etc.).


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@58366 aed309b6-a8cd-e111-996c-001c23d10e55