Revert sobomax's removal of --export-dynamic.  This is needed for ports like bugbuddy which require the signal handler symbols to be exported to libglade.  Other ports which require libglade may also benefit from --export-dynamic.
If you find GNOME 2.0 applications that complain libglade cannot find
a certain signal handler, recompile the port after updating to this
revision of glib20.  The problem should go away.

Note, this may not be the optimum fix, but it is what the GNOME people
probably intended, and gets things working for GNOME 2.0 Final.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@62226 aed309b6-a8cd-e111-996c-001c23d10e55