Added "threads" to the qmake.conf file. From this message on the list:
FreeBSD port installs threaded version of qt (qt-mt), but qmake.conf is
installed withouth thread in the CONFIG variable.  Because of that, qmake tried
to link with -lqt instead of -lqt-mt -pthread and fails.  With thread in the
CONFIG everything works.  This affects all ports that use qmake.

Folks, please test this (especially on current). It seems reasonable, is easy
to back out, I'm running a qt built with it right now.

Submitted by:	Tilman Linneweh <tilman@arved.de>


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@65308 aed309b6-a8cd-e111-996c-001c23d10e55