I agree with the Exim community that the exim-4.xx has reached production quality.
Move exim to exim-old for folks who need exim-3.xx, because the
configuration file for exim-4.xx is not backward compatible.  Move
exim-devel to exim, removing NO_LATEST_LINK:

        repo-copy       exim            -> exim-old
        copy over       exim-devel      -> exim
        retire          exim-devel

Slave ports are intended for use with the exim port, as before, so they
now build and install for exim-4.xx.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@66190 aed309b6-a8cd-e111-996c-001c23d10e55