Not using IGNORE anymore to avoid package building.
.if defined(BATCH) || defined(PACKAGE_BUILDING)
IGNORE= "You can not legally distribute binaries"
.endif

This was superfluous and inhibiting package builds of things that
depend on the port. Having RESTRICTED and NO_CDROM is enough to
ensure that a package will not appear on the FTP site or a CDROM
(it will be built and used as a basis for other packages to build
with, but will be deleted at the end of the build run).

Requested by:	kris
Reviewed by:	portmgr (silence)
PR:		42758


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@67730 aed309b6-a8cd-e111-996c-001c23d10e55