Change dependencies to depend on ghostscript-gnu instead of ghostscript-afpl. Bump PORTREVISION.
ocaml-images depends on ocaml-lablgtk,
which in the default configuration depends on some gnome components,
which depend on some other gnome components,
which depend on gnomeprint
which depends on ghostscript-gnu.
Because the two ghostscript versions conflict with each other,
This port failed on bento.

We probably need some bsd.port.mk magic to deal with the different
ghostscript versions.

Reported by:	kris


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@81151 aed309b6-a8cd-e111-996c-001c23d10e55