Add scummvm-tools.
This is a collection of various tools that may be useful to use in
conjunction with ScummVM. It includes tools that will allow you to
decompile and extract SCUMM scripts from various games, as well as
graphics and sound. Of special note is the extract utility, which
uses libmp3lame to compress a monster.sou file to about 25% its
original size.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@81962 aed309b6-a8cd-e111-996c-001c23d10e55