New port devel/p5-mixin
Mixin inheritance is an alternative to the usual
multiple-inheritance and solves the problem of knowing which
parent will be called. It also solves a number of tricky
problems like diamond inheritence.

The idea is to solve the same sets of problems which MI solves
without the problems of MI.

PR:		54746
Submitted by:	Mathieu Arnold <m@absolight.net>
Approved by:	fjoe (mentor)


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@85366 aed309b6-a8cd-e111-996c-001c23d10e55