All right, that's it, no more commits at 4am for me.
But I'm still going to leave it marked broken on all the other architectures.
That's because I think it really is broken on all the other architectures.
If it is not broken on all the other architectures, I am going to stick
my fingers in my ears and chant la-la-la-can't-hear-you.

arrrrrgh


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@91760 aed309b6-a8cd-e111-996c-001c23d10e55