Add p5-Class-ParmList 1.03, a module for processing named parameter lists for method calls.
This is a simple package for validating calling parameters to a subroutine
or method. It allows you to use "named parameters" while providing checking
for number and naming of parameters for verifying inputs are as expected and
meet any minimum requirements. It also allows the setting of default values
for the named parameters if omitted

PR:		58453
Submitted by:	clsung@dragon2.net


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@92019 aed309b6-a8cd-e111-996c-001c23d10e55