I specifically stated that I did not want to bump portepoch for this port.
I realize that my error in version numbering previously caused some confusion
about 9.2.3 being a more up to date version than 9.2.3.4, but this will quickly
be resolved with the next version, and affected only a few users who installed
the release candidate. The portepoch change is permanent, and perpetuates a
silly kludge for no good reason.

Please do not change this again without discussing it with me.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@92446 aed309b6-a8cd-e111-996c-001c23d10e55