Fix packaging when OSPFAPI is requested. This installs additional headers, and an additional directory.
Use the PLIST_SUB mechanism to achieve this rather than rewriting the entire
pkg-plist; if more sweeping additions are required we can do this later. The
smoke-and-mirrors going on here mean that we have to keep Makefile.inc around
in the WRKDIR until the install step, so that the pkg-plist can be substituted.

Keep the pkg-plist sort -r'd to avoid churn on future changes.

Noticed by:	kris


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@93958 aed309b6-a8cd-e111-996c-001c23d10e55