Add ocrad 0.5, OCR program implemented as filter.
GNU Ocrad is an OCR (Optical Character Recognition) program implemented
as a filter and based on a feature extraction method. It reads a bitmap
image in pbm format and outputs text in ISO-8859-1 (Latin-1) charset.
Also includes a layout analyser able to separate the columns or blocks
of text normally found on printed pages.
It can be used as a stand-alone console application, or as a backend to
other programs.

PR:		59232
Submitted by:	Ulrich Spoerlein <q@uni.de>


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@94543 aed309b6-a8cd-e111-996c-001c23d10e55