Fix the build on ia64:
The __libc_ia64_register_backing_store_base variable is defined on
Linux (in glibc) to allow processes to obtain the base of the RSE
backing store. On FreeBSD we do not have such a variable. We also
do not yet have a different interface for processes to use. So, for
now, hardcode the base address of the RSE backing store as it is
on FreeBSD. There's little chance this will change in the future,
so it's not that evil.

Approved by: portmgr (kris)


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@94803 aed309b6-a8cd-e111-996c-001c23d10e55