Bring the tix toolkit up to tcl/tk 8.4.
Also fix a broken shared lib configuration that caused the absolute
pathname (in the ports compilation directory) of the tix lib to be
recorded in the generated wish interpreter, so after removing the
compile directory from the machine, the interpreter could not be run
at all anymore.  Strange nobody ever comlained abou this.

Not Approved by:	maintainer (timeout)


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@94819 aed309b6-a8cd-e111-996c-001c23d10e55