Add mod_cvs-0.5.91 (known as mod_cvs2), apache2 version of mod_cvs.
mod_cvs recognizes if the files in your webtree are checked out
from a CVS repository. If so, it checks if the file is up-to-date
and, if not, updates it.
Another neat feature of mod_cvs is the date-checkout feature. You
can specify a date in the URI that gets passed to CVS, which checks
out an older revision of the requested file and sends it to the
user.

WWW: http://resare.com/noa/mod_cvs/

Approved by:    erwin (mentor) (implicitly)


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@97522 aed309b6-a8cd-e111-996c-001c23d10e55