Update to the 2004-01-14 snapshot of GCC 3.4.
Merge in my 2004/01/17 change to the gcc33 port to configure with
--program-suffix and related and further simplifications.

Merge in my 2004/01/13 change to the gcc33 port to make the automatic
generation of the package list for libraries and include files more
failure tolerant, so that at least `make install` now works on sparc64.

Merge in my 2004/01/05 change to the gcc33 port to combine and simplify
the post-install handling of target libraries and GCJ include files.

libgcj still is not supported and packaging is broken on sparc64; mark
BROKEN on that platform.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@98790 aed309b6-a8cd-e111-996c-001c23d10e55