Add linux-enemyterritory.
Wolfenstein: Enemy Territory is a full free multiplayer first person shooter
game. The game was originally going to be a retail expansion pack for Return
To Castle Wolfenstein but the project was cancelled and the good folks at
Activision decided to give it to us for free!

It's a team game; you will win or fall along with your comrades. The only way
to complete the objectives that lead to victory is by cooperation, with each
player covering their teammates and using their class special abilities in
concert with the others.

The game is unique in that it has moving objectives that makes the gameplay a
lot of fun!

WWW: http://www.enemy-territory.com

PR:		ports/61583
Submitted by:	Benjamin Lutz <benlutz@datacomm.ch>


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@98798 aed309b6-a8cd-e111-996c-001c23d10e55