Add latex-bibunits 2.2.20001010,
The `bibunits' package allows separate bibliographies for
different units or parts of the text. The units can be
chapters, sections or bibunit environments. The package is
compatible with a wide variety of packages, including, but not
limited to, natbib, overcite and KOMA-SCRIPT classes.

PR:		ports/61924
Submitted by:	hrs@FreeBSD.org


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@99191 aed309b6-a8cd-e111-996c-001c23d10e55