Fix the comment in the startup-script regarding how to enable clamav-milter in rc.conf.
Bumped PORTREVISION as this fix may be important for an user.

Mail is on its way to the maintainer but because of the actual virus
situation I do not wait for maintainer approval and commit the fix
right now (the submitter also submitted the startup-script for a previous
commit, so this is mainly a bugfix for a previously approved commit).

Submitted by:	Marius Strobl <marius@alchemy.franken.de>


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@99522 aed309b6-a8cd-e111-996c-001c23d10e55