Add mod_log_data 0.0.3, module for Apache 2.0 to log incoming and outgoing data.
mod_log_data is a module for Apache 2.0, logging incoming and outgoing
data.
This module can be used for debug or for security.
This beta version include one filter dumping the POST data for an
incoming request, and a filter dumping the outgoing data.

WWW: http://loghttp.sourceforge.net/

PR:		ports/57661


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@102209 aed309b6-a8cd-e111-996c-001c23d10e55