Unfortunately, the recent changes to math/xgfe (PR 59024) broke its compilation on FreeBSD 4 with GCC 2.95.4:         http://bento.freebsd.org/errorlogs/i386-4-latest/xgfe-2.1.log
The patch below fixes this.

During testing it on FreeBSD 4 and 5 with QT 3.3.1 I noticed that the
Advanced -> Multiple Files and Advanced -> Multiple Functions dialogs
were broken. I'm not sure if this is a new breakage with QT 3.3.1 or
if I didn't notice these when testing the previous patch on FreeBSD 5
with QT 3.2.1. Anyway, these are also fixed and as a precaution
PORTREVISION is bumped.

The patch also adds SIZE info.

PR:		64390
Submitted by:	Marius Strobl <marius@alchemy.franken.de>
Approved by:	maintainer


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@105715 aed309b6-a8cd-e111-996c-001c23d10e55