Add globus2,
The Globus Toolkit is an open source software toolkit used for
building grids. It is being developed by the Globus Alliance
and many others all over the world. A growing number of
projects and companies are using the Globus Toolkit to unlock
the potential of grids for their cause.

PR:		ports/65008
Submitted by:	Brooks Davis <brooks@FreeBSD.org>


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@105845 aed309b6-a8cd-e111-996c-001c23d10e55