Add p5-Class-DBI-Loader-Relationship 1.2, easier relationship specification in CDBI::L.
This module acts as a mix-in, adding the relationship method to
Class::DBI::Loader. Since Class::DBI::Loader knows how to map between
table names and class names, there ought to be no need to replicate
the names.  In addition, it is common (but not universal) to want
reverse relationships defined for has-many relationships, and for
has-a relationships to be defined for the linkages surrounding a
many-to-many table.

WWW: http://search.cpan.org/dist/Class-DBI-Loader-Relationship/

PR:		ports/65176
Submitted by:	Lars Thegler <lars@thegler.dk>


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@106499 aed309b6-a8cd-e111-996c-001c23d10e55