Ladies and Gentlemen,
It's wonderful to be here,
It's certainly a thrill.
You're such a lovely audience,
We'd like to take you home with us,
We'd love to take you home.
I don't really want to stop the show,
But I thought that you might like to know,
So let me introduce to you
The one and only Brooooooks Daaaaavis (brooks@)

Approved by:	portmgr GIs


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@108312 aed309b6-a8cd-e111-996c-001c23d10e55