Add rssh 2.1.1, a Restricted Secure SHell only for sftp or/and scp.
rssh is a Restricted Secure SHell that allow only the use of sftp or
scp. It could be use when you need an account (and a valid shell) in
order to execute sftp or scp but when you don't want to give the
possibility to log in to this user.

WWW: http://www.pizzashack.org/rssh/index.shtml

PR:		ports/65860
Submitted by:	enigmatyc


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@109638 aed309b6-a8cd-e111-996c-001c23d10e55