Add "fetchindex" target. It fetches the recent index file (INDEX _or_ INDEX-5).
A note about the implementation: I had to add the call to chmod to make
the index file readable to everyone. I'm a little bit surprised by this,
since my umask is 022. Either fetch(1) doesn't respect the umask or it
mirrors the permissions too (I haven't investigated this). Both
possibilities aren't documented in the man-page.

Approved by:	kris (previous version without the chmod)


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@109965 aed309b6-a8cd-e111-996c-001c23d10e55