Add xrsh, a port that contains two commands:
Xrsh is designed to allow you to start an X client on a remote machine
with the window displayed on the current server's $DISPLAY.  It has
many options that give you the ability to propagate environment
variables (including DISPLAY) to the remote system and works with
various types of X server access control including xauth and xhost.

Xrlogin opens a local xterm window and runs rlogin or telnet to
connect to a remote machine.

PR:		ports/67855
Submitted by:	KIMURA Yasuhiro <yasu@utahime.org>


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@112010 aed309b6-a8cd-e111-996c-001c23d10e55