Plug a big memleak in the mimedefang embedded perl mode. The fix seems to be FreeBSD specific, other platforms (linux,solaris) have other problems (stability).
The global variable "PL_perl_destruct_level" needs to be set to "1" before
the destructor perl_destruct() is called, not afterwards. It is not clear why
this has to be done, since "PL_perl_destruct_level" should be defined static.
Without this fix "PL_perl_destruct_level" is always "0" when the destructor
gets called.

This fix has an effect only if "MX_EMBED_PERL=yes" is set in the startup-
script which is needed to run mimedefang in the embedded perl mode. You
should use perl 5.8.4 too to have it working properly.

The fix has been committed to mimedefang 2.44-BETA-1.

Reviewed by:		David F. Skoll <dfs@roaringpenguin.com>
Not reviewed by:	Maintainer (no reply)


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@112569 aed309b6-a8cd-e111-996c-001c23d10e55