Add arkpandora 2.04,
This font pack is designed for screen and print and as a
"drop-in" replacement for Arial and Time New Roman, so people
can stop using not-so-free Webfonts from Microsoft.

You get the following fonts: Aerial, Aerial Mono, Tymes &
Veranda (no this is not a typo).

It will automatically make the fonts recognizable by the
system.  It also makes Aerial the default font for the
Sans-Serif and Tymes the Serif font types.

The fonts are derived from the Bitstream Vera fonts and are
available under the same terms as Vera.

PR:		ports/70564
Submitted by:	Alexey Dokuchaev <danfe@regency.nsu.ru>


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@116547 aed309b6-a8cd-e111-996c-001c23d10e55