Take over maintainership from demon.
Work around an incompatibility between fltk (a dependency) and libXft
version 2 which results in flwm getting a SIGSEGV at startup. This
basically prevents the vertical labels from being drawn in the vertical
title bar on the left of the window. Otherwise there is no change in
functionality.

I tried using newer versions of fltk but they either had the same problem
or were so new that flwm could not be linked.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@117460 aed309b6-a8cd-e111-996c-001c23d10e55