Add enblend
Enblend is a tool for compositing images. Given a set of images that
overlap in some irregular way, Enblend overlays them in such a way that
the seam between the images is invisible, or at least very difficult to
see. Enblend does not line up the images for you. Use a tool like Hugin
to do that.

PR:		ports/71992
Submitted by:	cartola@openit.com.br
Approved by:	adamw (mentor)


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@121832 aed309b6-a8cd-e111-996c-001c23d10e55