GQ changed the configure option to specify the LDAP base directory from --with-ldap to --with-ldap-prefix quite some time ago.
Incidentally, it still used to work with the wrong option most of the
time, as configure silently ignores unknown --with options (this is a
feature), and as the configure script often included the correct -I
and -L options to find OpenLDAP as a side-effect of other stuff
installed in the same location.  However, I've seen it fail as well,
and using the correct option is the canonical way out.

Approved by:	maintainer timeout


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@123301 aed309b6-a8cd-e111-996c-001c23d10e55