Add adstudio 4.0, a database query and administration tool.
Aqua Data Studio is a database query and administration tool that allows
developers to easily create, edit, and execute SQL scripts, as well as
browse and visually modify database structures. Aqua Data Studio provides
an integrated database environment with a single consistent interface to
all major relational databases.

It is free for personal or educational use, but a commercial license is
required for all other usages.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@128590 aed309b6-a8cd-e111-996c-001c23d10e55