- Add cactid 0.8.6d
Cactid is a poller for Cacti that primarily strives to be as fast as
possible. For this reason it is written in native C, makes use of POSIX
threads, and is linked directly against the net-snmp library for
minmumal SNMP polling overhead. Cactid is a replacement for the default
cmd.php poller so you must decide if using Cactid makes sense for your
installation.

WWW: http://www.cacti.net/cactid_info.php


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@129856 aed309b6-a8cd-e111-996c-001c23d10e55