New port: devel/pear-Event_Dispatcher PEAR dispatch notifications using PHP callbacks
The Event_Dispatcher acts as a notification dispatch table.
It is used to notify other objects of interesting things. This
information is encapsulated in Event_Notification objects. Client
objects register themselves with the Event_Dispatcher as observers of
specific notifications posted by other objects. When an event occurs,
an object posts an appropriate notification to the Event_Dispatcher.
The Event_Dispatcher dispatches a message to each registered
observer, passing the notification as the sole argument.

PR:		ports/78889
Submitted by:	Antonio Carlos Venancio Junior <antonio@php.net>


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@131454 aed309b6-a8cd-e111-996c-001c23d10e55