Add sidplayer 4.4,
SIDPlayer is a replayer program for C64 music (SID tunes). You can
think of it as being a stripped-down C64 emulator that only emulates
the processor and the sound chip of the C64. It also supports the
SID option of the Catweasel MK3 PCI card.

PR:		ports/81537
Submitted by:	Emanuel Haupt <ehaupt@critical.ch>


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@136256 aed309b6-a8cd-e111-996c-001c23d10e55