New port: Xfce Foundation Classes.
The Xfce Foundation Classes (XFC) is a set of well integrated C++ classes
for developing Xfce applications on UNIX-like operating systems.

XFC combines the power of GTK+ and the power of C++ into a state-of-the-art
application development framework for the Xfce Desktop Environment.  XFC
judiciously uses C++ language features to avoid layering on too much extra
C++ complexity.  Its API is easy to understand and use, and should feel
immediately familiar to most GTK+ programmers.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@136625 aed309b6-a8cd-e111-996c-001c23d10e55