Python ICQ-Transport for Jabber.
The purpose of this transport is to provide a way for users to be logged into
ICQ via their Jabber account. It is implemented in Python, using the Twisted
framework. This project is just getting off the ground, and so it does not have
quite the functionality of other transports yet. That will come in time.

PR:		ports/81994
Submitted by:	Renato Botelho <freebsd@galle.com.br>


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@136996 aed309b6-a8cd-e111-996c-001c23d10e55