Assume maintainership.
It's important to note here that these particular versions of autotools
will be left without FreeBSD-specific hacks, so that developers and IDEs
alike will be able to run in a pristine environment, ensuring
cross-platform capability.  The prefix of ${LOCALBASE}/gnu-autotools
will remain untouched, it's merely a case of ensuring that they are
kept up to date.

Approved by:	marcus


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@137523 aed309b6-a8cd-e111-996c-001c23d10e55