The tochnog port installs some plugins into a directory which is owned by the the cad/linux-gid port, but it doesn't adds a dependency to linux-gid, since tochnog can be used standalone.
The right thing to do would be to add a slave port which installs the
plugins only. The slave port would have to depend upon linux-gid and
tochnog.

Since the linux-gid port installs now in X11BASE and tochnog installs in
LOCALBASE, we have a plist problem ATM too.

In light of the recent discussion about the right base directory for ports
which use X11 bits I decided to go the quick route until everything is
decided:
 - move linux-gid back to LOCALBASE
 - let tochnog still install the plugins

This reverts back to the status quo (a working tochnog/linux-gid). Someone
with a good asbestos suite may want to mark the tochnog port BROKEN.

Dependencies explained by:	pfgshield-freebsd@yahoo.com


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@138033 aed309b6-a8cd-e111-996c-001c23d10e55