Add linux-ut2003-demo.
Unreal Tournament 2003 is a first-person shooter computer game designed mainly
for multiplayer gaming although the game had a built in single-player mode
that mimics multiplayer gaming by featuring AI-bots. The game is part of the
Unreal franchise's series of games and is a the sequel to 1999's Unreal
Tournament. Unreal II: The Awakening was released as a sister product to the
game, however, was developed for single-player only. Later, Unreal II would
receive an add-on that would enable a multiplayer mode. UT 2003 was followed
by Unreal Tournament 2004 released in March of 2004.

WWW: http://www.unrealtournament.com/ut2003/

PR:		ports/85549
Submitted by:	Alejandro Pulver <alejandro@varnet.biz>


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@141860 aed309b6-a8cd-e111-996c-001c23d10e55