Fix allocation of pty for shell-mode.
This was broken for newer FreeBSDs which give out pty names
/dev/pty[pqrsPQRS][0123456789abcdefghijklmnopqrstuv]

emacs, manually searching, stopped at a gap of 3 entries.  But the new
naming scheme allows a gap of 28.  Set to 30 just in case.  Actually
there's no good reason not to set that to 256, it's not that 512
system calls kill us everytime someone opens a shell in emacs.

Bump portsrevision so that it gets picked up by portupgrade.

I wonder whether other emacsens are affected, too...


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@143438 aed309b6-a8cd-e111-996c-001c23d10e55