New port
This is a mini-SNMP daemon. The basic daemon implements the system group
and a number of private extensions to manage the UDP transport mapping,
communities, trap destinations and loadable modules. In this form it can
be used to provide remote access to arbitrary data that can be described in
the form as required by the SMI. The daemon speaks both SNMPv1 and SNMPv2c.

PR:		86400
Submitted by:	Andrey V. Elsukov <bu7cher@yandex.ru>
Approved by:	pav (mentor)


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@143637 aed309b6-a8cd-e111-996c-001c23d10e55