Pachi el marciano is a platforms game inspired by games like Manic Miner and Jet Set Willy from the legend Mathew Smith.
The goal of the game is to collect all the objects of each level, when this is
made, the exit gate to the next stage will appear and you can pass trough.

Pachi is a comic character created by Nicolas Radeff in 2001.

PR:		86845
Submitted by:	Dmitry Marakasov <amdmi3@mail.ru>
Approved by:	novel (mentor)


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@144391 aed309b6-a8cd-e111-996c-001c23d10e55