Add support for use_md_swap.  When this variable is set in a portbuild.conf, builds will each be done in a separate swap-backed md. This dramatically improves build performance since
* Every transaction is not written to disk, so disk bandwidth is not
  a bottleneck

* Multiple builds do not contend with each other for the same set of
  filesystem locks and other per-device resources

The size of the md devices is controlled by the md_size variable.  '2g'
seems to be a good size.

Currently we mdconfig -u each device after each port build, since
otherwise dirty blocks accumulate and the md eventually uses a full
amount of backing store (2g in the above example).  This is a problem
if there is unsufficient swap backing to accomodate them all.

XXX This should be made configurable to avoid the performance penalty on
systems that do have enough swap backing


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@144892 aed309b6-a8cd-e111-996c-001c23d10e55