Update the antiquated version of the GNAT compiler to the GPL-2005 version.
Note that this version now produces executables that are
covered by the GPL as opposed to the previous version which was
covered by the modified GPL.

Also note that the GNAT sources in the stock GCC-4 tree are
not under this restriction.  A new port based on the sources
in GCC-4 will come when I get the time.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@149845 aed309b6-a8cd-e111-996c-001c23d10e55