This is a collection of greek TTF fonts licensed under GPL.
These fonts are designed for the Greek language and are iso-8859-7 compatible.
They are saved though in Unicode format, to be compatible with modern fonts and
in order to be easily expandable.  It consist of the following families:

a) Bonati
c) Jacobs
b) Optima
d) OptimaCollege
e) ParisAifel

PR:		ports/89978
Submitted by:	Panagiotis Astithas <past@ebs.gr>


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@150827 aed309b6-a8cd-e111-996c-001c23d10e55