Update to the 20051220 snapshot of GCC 3.4.6.
Install the .info files of the lang/gcc34 port in a port-specific
subdirectory, which allows us to remove the conflicts with lang/gcc40
and lang/gcc41 and also removes the hidden conflict with lang/gcc33.

Also, convert pkg-plist to use a new substitution (%%SUFFIX%%) instead
of hardcoding the version number 34.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@151805 aed309b6-a8cd-e111-996c-001c23d10e55