Update to the 20051215 snapshot of GCC 4.0.3.
Install the .info files of the lang/gcc40 port in a port-specific
subdirectory, and move include/mf-runtime.h into a version specific
directory.   This allows us to remove the conflicts with lang/gcc33,
lang/gcc41 and lang/gcc42.

Also, convert pkg-plist to use a new substitution (%%SUFFIX%%) instead
of hardcoding the version number 40.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@151809 aed309b6-a8cd-e111-996c-001c23d10e55