Now that new style rc.d scripts are being run as part of the base rcorder, hard coded variable values in these scripts are overriding the values in /etc/rc.conf[.local] (due to the way that variables from the latter are read at boot time).
Therefore, change the boot scripts to set default values only
if the variable is unset in /etc/rc.conf[.local]. This will
allow the service to start at boot time if it's been enabled
as the user would expect.

This change will be a noop for users who have systems that
have not yet been upgraded to the new rc.d code in the base.

In many cases there are other variables in the scripts that
should get similar treatment, however I did not change
anything other than the _enable lines. I'll leave the rest
up to the maintainers to do as they see fit.

Bump PORTREVISION to make sure that users and packages
pick up this change.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@152963 aed309b6-a8cd-e111-996c-001c23d10e55