Add www/mod_pubcookie:
Pubcookie is an open-source package for intra-institutional
single-sign-on end-user web authentication. More generally, it is an
approach to identifying users as they browse to an institution's many
websites that require authentication. It helps an institution reuse
existing authentication services (like Kerberos, LDAP, or NIS), and it
limits the exposure of end-user passwords by ensuring they're only sent
to a trusted login service.

This port contains the apache module portion of the Pubcookie system.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@155471 aed309b6-a8cd-e111-996c-001c23d10e55