Repetitive Strain Injury is an illness which can occur as a result of working with a mouse and keyboard. This utility can be used to remind you to take a break now and then. It will show you a random picture from a collection you can configure yourself for a configurable duration at a configurable interval.
You can use these breaks to do some stretch exercises for example, or as
a reminder to walk away from the computer for a while.

RSIBreak will sit in your system tray and when it is time for a break it
will show you the picture full screen. All timings can be set by clicking
with the right mouse button on the icon in the system tray.

WWW: http://www.rsibreak.org/


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@157136 aed309b6-a8cd-e111-996c-001c23d10e55