Update to the 20060316 snapshot of GCC 4.0.4.
Now that GCC 4.1.0 has been released, remove support for the Fortran
frontend which was still quite immature in the 4.0 series compared to
what we now have in the the 4.1 series.

Always build both shared and static libraries instead of having these as
two exclusive options defaulting to the former.

Do not strip some of the the binaries any longer.

No longer hardcode the version number in LATEST_LINK.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@157494 aed309b6-a8cd-e111-996c-001c23d10e55