Do not run-depend on festival -- it really ought to be the other way around.
For festlex-cmu use the 2004 version found in the 1.95 directory, so that
it works with audio/festival. The others did not change since 1999.

Stop the silly practice of extracting into ${WRKDIR} and copying into
destination -- extract directly.

Replace @dirrm with @dirrmtry in a few places and unbreak by adding more
of same.

The same should be applied to other audio/fest* ports, but these are
enough to get the English-only speech out of festival.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@158932 aed309b6-a8cd-e111-996c-001c23d10e55