Mars, Land of No Mercy is a turn-based strategy game setting on Mars during the early stages of human colonization.The player embodies the leader of a mercenary team, landed on Mars to take advantage of conqueror battles between the Worldwide Colonizer Corporations.
The main purpose is to command his team trying to find them commissions,
training and leading them in battle, while still being aware of financial and
instrumental resources by administering them.The Mech is the fundamental craft
used by mercenaries, but they won't lack of other kind of units, as well as
troops.

The game most remarkable point is the incredible liberty given to the player
and therefore the great strategic side that derives. It is indeed up to the
player to decide how to behave towards the one who recruited him, and the
enemies of turn.

Basically every action leads to a reaction, so it is strongly recommended to
keep a good and consistent behavior to achieve confidence from Corporations,
and avoid being excluded by them.

WWW:	http://mars.sourceforge.net/

PR:		ports/94771
Submitted by:	Jose Alonso Cardenas Marquez <acardenas@bsd.org.pe>


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@159175 aed309b6-a8cd-e111-996c-001c23d10e55