Remove broken extra patch files/extra-patch-module-mod_ldap.c the patch itself is actually repeated 3 times so it tries to reverse itself and fails. This patch is only applied if the LDAP know is checked so it doesn't appear on all builds.
The patch itself defines TLS support in the LDAP module. There are two problems
with doing this. One, it breaks LDAP if the server doesn't support TLS. And two
it causes the build to fail on 4.x. The patch to do this was a requested option,
and isn't critical for most people.

The maintainer is going to add another option knob for LDAP_TLS support once the
4.x build problems are resolved.

PR:		96713
Submitted by:	Beech Rintoul <beech@mangohealth.org> (maintainer)


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@162025 aed309b6-a8cd-e111-996c-001c23d10e55