Add port www/xpi-resurrectpages:
Dead pages, broken links, the scourge of the internet. Powerhouse
sites like Slashdot and Digg can bring a server to its knees. What
do we do when a page is dead but we still want to see it? Call in
the clerics, and perform a resurrection ceremony! Or, the easier
route, use this extension.

WWW: http://www.arantius.com/article/resurrect

Approved by:	krion (mentor)


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@163595 aed309b6-a8cd-e111-996c-001c23d10e55