Retrieve list of filesystems and their properties
Sys::Filesystem is intended to be a portable interface to list and query
filesystem names and their properties. At the time of writing there were
only Solaris and Win32 modules available on CPAN to perform this kind of
operation. This module hopes to provide a consistant API to list all,
mounted, unmounted and special filesystems on a system, and query as
many properties as possible with common aliases wherever possible.

WWW: http://search.cpan.org/dist/Sys-Filesystem/

PR:		ports/98307
Submitted by:	pirzyk


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@166572 aed309b6-a8cd-e111-996c-001c23d10e55