The configure script distributed with version 0.7 of yersinia checks for multiple pcap.h files and aborts with an error message, if more than one is actually present. This appears to be a pre-caution to prevent building with pcap.h from one version and running with a library belonging to another one ...
While I see the value of such a test, it prevents building on systems
before 5.0 (where it requires a port version of libpcap). After some
testing I'm quite confident, that the port will build correctly (i.e.
without version mismatch between header and library), if the port
version of libpcap is present.

The libpcap port is used instead of the system version, if present.
In that case, yersinia is statically linked with libpcap.

The problem was detected by jau at iki dot fi (Jukka A. Ukkonen).

PR:		99869


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@167363 aed309b6-a8cd-e111-996c-001c23d10e55