A plugin for the Trac project/source code management system that creates Gantt charts based on ticket reports. It is a work-in-progress.
A Gantt chart is a graphical representation of the duration of tasks
against the progression of time. Gantt charts are very useful for
planning a scheduling a project based on different tasks that will
need to be completed and their durations and dependencies on other
tasks.

The TracGantt plugin uses Trac's ticketing system to generate Gantt
charts based on existing Trac Ticket Reports (found under the "View
Tickets" tab in Trac). The Gantt charts are pure XHTML and are
generated on-the-fly from the ticket reports, using only the tickets
provided by the report.

WWW: http://ideas.water-powered.com/gantt/#docs

PR:		ports/100140
Submitted by:	Dryice Liu <dryice at dryice.name>


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@167681 aed309b6-a8cd-e111-996c-001c23d10e55