boxquote provides a set of functions for using a text quoting style that partially boxes in the left hand side of an area of text, such a marking style might be used to show externally included text or example code.
,----
| The default style looks like this.
`----

A number of functions are provided for quoting a region, a buffer, a
paragraph and a defun. There are also functions for quoting text while
pulling it in, either by inserting the contents of another file or by
yanking text into the current buffer.

WWW: http://www.davep.org/emacs/#boxquote.el


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@167788 aed309b6-a8cd-e111-996c-001c23d10e55