Add port www/linux-mplayer-plugin:
- Introduce NPAPI infrastructure and a demo port, both at an alpha stage

mplayerplug-in is a browser plugin that uses mplayer to play digital
media from websites.

This plug-in gives Mozilla the ability to play media from a website
the net without reading the source html and getting the url manually.
Media is played embedded in the page or in a separate window depending
on how the author of the webpage intended the media to be seen.

WWW: http://mplayerplug-in.sourceforge.net/

Wiki:		http://wiki.freebsd.org/NPAPI


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@168976 aed309b6-a8cd-e111-996c-001c23d10e55