Capture the Flag offers an exciting variation on team play. It gives more meaning to base defense along with offensive positioning and organized strikes.
The basic premise is this:

Attack the enemy base, grab their flag, then take it back to your base. In
this variation you must touch your base flag when carrying the enemy flag in
order to score - if your flag isn't at your base when you return, you or your
team must get it back to gain points and captures.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@169009 aed309b6-a8cd-e111-996c-001c23d10e55