Update to snapshot 20061102.
* The volume probing helper can take up to a minute to run the isoinfo
  command against data DVDs, so bump the timeout up to 60 seconds.

* Fix a spurious warning when running lshal

Add two new properties to SCSI block devices.  The first is
block.freebsd.cam_path which is a bus,target,lun path to a SCSI device.
This is needed by tools such as cdrecord, and by adding a private freebsd
property, it is much easier to get apps like nautilus-cd-burner working.

The second property is block.freebsd.atapi_cam_device.  This property is
actually attached to ATAPI CAM devices, and denotes the CAM block device
associated with this ATAPI block device.  This may be useful for programs
that can only operate with CAM-compliant devices.

Finally, update the volume probing code so that blank discs are treated as
volumes.  This is sort of crude.  We basically say a disc is blank is it
is a disc and contains no data and no audio.

With all of these changes, plus a minor change to nautilus-cd-burner, n-c-b
now works again on FreeBSD.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@176243 aed309b6-a8cd-e111-996c-001c23d10e55