Update to 9.10, see changelogs for details:
	http://www.opera.com/docs/changelogs/linux/910/

The most important part in this release is added a new feature, fraud
protection. It is a phishing filter and keep in mind that it is not enable by
default. To enable it, go to Tools > Preferences > Advanced > Security and
check the "Enable Fraud Protection". I am using it right now and it doesn't
affect any speed. To learn more about fraud protection, visit here:

	http://www.opera.com/docs/fraudprotection/


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@180101 aed309b6-a8cd-e111-996c-001c23d10e55