Fix a compiler warning in the AVR-COFF generator that appeared with GCC 4.x, and caused the compilation to fail due to the -Werror option used.
Remove an annoying assembler warning in the ATmega256x patch that was
frequently triggered by large C switch() statements in GCC, and
generally considered to be a regression.  The author of that patch
agreed that the warning isn't particularly useful anyway for a number
of reasons, and recommended to just drop it.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@181423 aed309b6-a8cd-e111-996c-001c23d10e55