New port linux-ov511-kmod, a usb webcam driver based on the linux driver 'ov511' (from the linux sources) built on FreeBSD using devel/linux-kmod-compat. The driver supports some ov511 based webcams.
THIS IS EXPERIMENTAL CODE, TRY IT AT YOUR OWN RISK.

This is only for 6.x/7.x and at the moment only for i386.

Most bugs and limitations are due to linux-kmod-compat rather than
this port itself, so upgrading linux-kmod-compat (and rebuilding
this port) will most likely improve functionality and remove some bugs.

As usual if some of the port gurus wants to help fixing any mistakes
that i may have made in the port (or missing things, e.g. a suitable
warning message when the port is installed), i will be very grateful
(and can spend more time working on the emulation code).

In particular, i could not find a good site (ftp or web) to fetch
individual linux source files from, so i copied them on my web server
and used it as MASTER_SITE. This should be fixed though.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@183971 aed309b6-a8cd-e111-996c-001c23d10e55