M.A.M.E. Catalog / Launcher II - also referred to as QMC2 - is the successor of one of the first XMAME GUI frontends available on this planet called QMamecat (derived from MAMECAT, which was text-only).
QMamecat was based upon Qt 2; its development was frozen in 2003. QMC2 has
been rebuilt from scratch as a Qt 4 project. Parts of the design and code were
inspired by its predecessor, but it's not just a remake. We tried to make the
new design as flexible as possible to minimize dependencies from frontend- and
CLI-related MAME changes, which was a major deficiency of QMamecat. QMC2 now
uses a template based MAME configuration scheme, which can easily be enhanced
with additional command line options (defined in an XML template file).

As a result of this flexible design, we are able to support multiple MAME
targets - currently XMAME and SDLMAME!

WWW: http://www.mameworld.net/mamecat/


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@186179 aed309b6-a8cd-e111-996c-001c23d10e55