FreeCAD is an OpenSource CAD/CAE, based on OpenCasCade, QT and Python.
It features some key concepts like macro recording, workbenches, ability to run
as a server and dynamically loadable application extensions and it is designed
to be platform independent.

Warning: FreeCAD is still in ALPHA state and not in shape for end user usage!

<http://juergen-riegel.net/FreeCAD/Docu/>

Suggested by:	Pedro F. Giffuni <giffunip (at) yahoo.com>


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@191927 aed309b6-a8cd-e111-996c-001c23d10e55