- Make portconf work when /usr/ports is a symlink.  When the link is changed,   you have to reinstall portconf to get make.conf right.
As this does not affect most people, PORTREVISION is not bumped.

The patch in the PR does not work: we can not use variable in substitution.
So, the only way to do is to detect the real path of /usr/ports at
installation time.

PR:   	      	ports/104838
Submitted by:	rafan
Approved by:	maintainer timeout (8 months)


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@194058 aed309b6-a8cd-e111-996c-001c23d10e55