Eric is a full featured Python and Ruby editor and IDE, written in python.
It is based on the cross platform Qt gui toolkit, integrating the highly
flexible Scintilla editor control. It is designed to be usable as everdays'
quick and dirty editor as well as being usable as a professional project
management tool integrating many advanced features Python offers
the professional coder.

This is a port of eric4 (based on Qt4).

WWW: http://www.die-offenbachs.de/detlev/eric.html


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@196287 aed309b6-a8cd-e111-996c-001c23d10e55