Add port games/windstille:
In Windstille the player will be placed in a foreign alien world
and has to find its way back into safety. Special focus will be on
getting the foreign world and the players character itself believable.
Exploration will be more needed than plain shoot action, even so
the players-character will be equipped with a multifunctional
war-suit, so he will be far from defenseless. The game will have
slow placed explore and seak sequences as well as more rough section
that will feature fast shoot and jump action. The game world should
be presented in a consistent state, enemies once destroyed should
not respawn automatically. The player is free to go forward and
backward in the world as long as the level design allows it.

WWW: http://windstille.berlios.de

PR:		ports/113693 (based on)
Submitted by:	Dmitry Marakasov <amdmi3@amdmi3.ru>


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@196983 aed309b6-a8cd-e111-996c-001c23d10e55