Add port games/kardsgt:
KardsGT is a free sourced card game suite. We want to bring an
immersing card game environment to the free sourced world. It is
currently in a beta-level of development. Despite that it is quite
playable. Currently you can play Cribbage, Euchre, Hearts, Old Maid,
Spades, and War.

WWW: http://kardsgt.nongnu.org/


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@197317 aed309b6-a8cd-e111-996c-001c23d10e55