Add new port: irc/irchat-jp
Irchat-jp is an IRC client written in Emacs Lisp, derived from irchat.
This is one of the oldest IRC client that handles ISO-2022-JP character
encoding (another one is ports/japanese/ircII).

Please note that this is irchat-JP, not irchat-PJ that is already in
FreeBSD ports repository (ports/irc/irchat-pj-*).

Reviewed by:    kuriyama


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@197335 aed309b6-a8cd-e111-996c-001c23d10e55