Add port x11/xsel-conrad:
XSel is a command-line program for getting and setting the contents of
the X selection. Normally this is only accessible by manually
highlighting information and pasting it with the middle mouse button.

This port is similar to x11/xsel, but with different CLI syntax and
a bit more functionality. It is a lot more popular, too.

WWW: http://www.vergenet.net/~conrad/software/xsel/
Author: Conrad Parker <conrad@vergenet.net>

Based on:	OpenBSD port, Portage package


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@197386 aed309b6-a8cd-e111-996c-001c23d10e55