Add port databases/tablelog:
Tablelog is a contributed PostgreSQL module that uses a trigger to log
any INSERTs, UPDATEs and DELETEs on a specific table into another table.
The second part of tablelog is able to restore the state of the original
table or of a specific row for any time in the past.

WWW: http://tablelog.projects.postgresql.org/
Author: Andreas Scherbaum <ads-pgfoundry @nospam@ wars-nicht.de>

PR:		ports/114541
Submitted by:	Robert Gogolok <gogo@cs.uni-sb.de>


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@197772 aed309b6-a8cd-e111-996c-001c23d10e55