Re-structure support for libraries in GHC: This commit splits off the supporting libraries that pull in additional dependencies like X11 and GL. These will be provided by separate ports like x11/hs-x11-ghc using the "Cabal"-mechanism (a framework for installing additional libraries into Haskell environments like GHC and NHC). Accordingly, the libraries will put their files into a subdirectory of GHC. List of libraries that have been split off: ALUT GLUT HGL OpenAL OpenGL X11
Also, finally remove dependency on devel/readline: the current readline in
base has all the features we need.

Bump PORTREVISION accordingly.

Tested through Tinderbox runs on lang/ghc and devel/darcs. Please be patient
for the supporting libraries to roll in in the next few days. All screwups
are my fault only and should not be attributed to haskell@.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@199587 aed309b6-a8cd-e111-996c-001c23d10e55