Update to version 1.5. There are many changes to improve portability to other platforms (notably to Linux) and several substantive changes:
Bug fixes
=========
1. Add error handling for failed TDIR creation in ppf_decrypt,
ppf_verify, and ppf_mime*.
2. Fix error handling in ppf_mime_decrypt.

Improvements
============
1. When displaying MIME messages after verification, substitute a space
for the MIME'ified =20 character at the end of a line, and substitute
an = sign for the MIME'ified =3D. This greatly improves readability
of the verified message.
2. Add a --- Status --- line for the beginning and end of the PGP
stuff which includes a timestamp to help avoid a socially engineered
message that includes the "good signature" output in the same format
as what we print.

Notes
=====
1. Support for other than GnuPG was dropped in this version, which I
don't think will actually affect anyone.
2. I did do a version 1.4, but it was a limited release.

See CHANGES for more information.

Approved by:	portmgr (erwin)


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@202874 aed309b6-a8cd-e111-996c-001c23d10e55