Upgrade to suPHP 0.6.2.  This upgrade is quite major; it introduces numerous features necessary for hosting, adds the use of a configuration file (vs. compile-time settings) for configuration details, and is presently supported by the software author.
Changes to httpd.conf are required for this to work, otherwise PHP
content is offered for download and not parsed.  pkg-message contains
full migration details, but here's the simple version:

suPHP 0.5.2                            suPHP 0.6.2
===========                            ===========
AddType application/x-httpd-php .php   AddType x-httpd-php .php
AddHandler x-httpd-php .php            suPHP_AddHandler x-httpd-php

Tested on RELENG_6 and RELENG_7, using Apache 2.0 and 2.2, with the
help of the PR submitter.

Testing for Apache 1.3 was not performed, thus the port requires the
use of 2.0 or later.  suPHP 0.6.2 does support Apache 1.3, but I'll
need someone running 1.3 to work with me on fixing the port to support
it, if desired.

PR:		ports/116281
Submitted by:	Vladimir Korkodinov <viper@perm.raid.ru>


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@202998 aed309b6-a8cd-e111-996c-001c23d10e55