This takes from a search and turns it into a hash.
The returned has is in the following format.

{DN}{ldap}{attribute}[array of values for this attribute]

The reason for the {ldap} is to allow for other values and
the like to be tagged onto a hash for a DN that are unrelated to LDAP.

This function does not make any attempt to check if the search succedded
or not.

WWW: http://search.cpan.org/dist/Net-LDAP-LDAPhash/

PR:		ports/120589
Submitted by:	Zane C, Bowers <vvelox at vvelox.net>


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@207456 aed309b6-a8cd-e111-996c-001c23d10e55