- add p5-LWP-UserAgent-Determined-1.03
This class works just like LWP::UserAgent (and is based on it, by being a
subclass of it), except that when you use it to get a web page but run into a
possibly-temporary error (like a DNS lookup timeout), it'll wait a few seconds
and retry a few times.

It also adds some methods for controlling exactly what errors are considered
retry-worthy and how many times to wait and for how many seconds, but normally
you needn't bother about these, as the default settings are relatively sane.

WWW:    http://search.cpan.org/dist/LWP-UserAgent-Determined/


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@209177 aed309b6-a8cd-e111-996c-001c23d10e55