Update to 8.2.7
Release notes:
http://developer.postgresql.org/pgdocs/postgres/release-8-2-7.html

Remove HIER patch option, since it is not available anymore. The site
is down and original author responses that he does not maintain it
anymore.

PR:	ports/116423
Remove option to build and link with thread safety, and always build
thread safe instead. Decided after a discussion with postgresql
hackers, and tested for a while in postgresql83. This helps when
for example linking with openldap, or for using certain programming
languages as postgresql functions.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@209327 aed309b6-a8cd-e111-996c-001c23d10e55