New Implementation ================== The previous code tried to handle moved ports in much the same way as it handled replacing a port with the -o option, but this would sometimes create problems if we are deep in a dependency graph. So rather than re-implement only a portion of the -o code, use the whole thing instead.
We accomplish this by either exec'ing ourselves with the proper -o
options if we are in the dependency tree, or by asking the user to
restart with the proper -o options if they specified the moved port
on the command line. The latter is necessary to avoid having to
duplicate all the "initialize the parent process" code.  This allows
me to simplify the code in a few places since -o is now the only
special case.

Bug Fixes
=========
1. Don't try to derive $portdir from $upg_port if using -o
2. Don't try to build a port with an +IGNOREME file if we are doing
the dependent ports in -r mode

Minor tweak: Add some whitespace after a warning message


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@209402 aed309b6-a8cd-e111-996c-001c23d10e55