WIP of the NMM port (to allow further, distributed enhancements), see below for what NMM is.
Status so far:
 - compiles and installs (KDE integration should be possible with this)
 - doesn't segfault when doing a ruidemental test
 - doesn't provide the complete plugin set (no RTP, no shoutcast,
   no OSS audio (any volunteers to develop this?), ...)
 - TODO list integrated into the port Makefile

From pkg-descr:
---snip---
The Network-Integrated Multimedia Middleware (NMM) provides an architecture
that allows to create distributed multimedia application easily: local and
remote multimedia devices or software components can be controlled
transparently and integrated into a common multimedia processing flow graph.

As an example, this allows for the quick and easy development of an
application that receives TV from a remote device -- including the
transparent control of the distributed TV receiver. Even a PDA with only
limited computational power can run such an application: the media
conversions needed to adapt the audio and video content to the resources
provided by the PDA can be distributed within the network. While the
distribution is transparent for developers, no overhead is added to all
locally operating parts of the application.

WWW: http://www.motama.com/
---snip---


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@211765 aed309b6-a8cd-e111-996c-001c23d10e55