Add tclxml-expat , API for parsing XML documents using Tcl.
TclXML is an API for parsing XML documents using the Tcl scripting
language. It is also a package with several parser implementations.
The goal of the TclXML package is to provide an API for Tcl scripts
that allows "Plug-and-Play" parser implementations; ie. an application
will be able to use different parser implementations without change
to the application code.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@212661 aed309b6-a8cd-e111-996c-001c23d10e55