Add httpry:
"httpry is a specialized packet sniffer designed for displaying and logging
HTTP traffic. It is not intended to perform analysis itself, but to capture,
parse, and log the traffic for later analysis. It can be run in real-time
displaying the traffic as it is parsed, or as a daemon process that logs to an
output file. It is written to be as lightweight and flexible as possible, so
that it can be easily adaptable to different applications."

WWW: http://dumpsterventures.com/jason/httpry/

Submitted by:	Richard Bejtlich (private mail)


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@214799 aed309b6-a8cd-e111-996c-001c23d10e55