Update WWW and MASTER_SITES URLs.
The main master site does not actually have this particular version of
gringotts, but it seems better to list the new URL as the old stopped
working long time ago. A new version of gringotts exists, and if anyone
is using and have an interest in this port they should look into it.
Keeping the secondary master site so that there is at least one place
to fetch the current distfile.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@218383 aed309b6-a8cd-e111-996c-001c23d10e55