The great "Unbreak archivers/lzo2" project - step 1
This patch doesn't unbreak it yet, but it increases the output of
the test in which the port building hangs. I tried to build this
port on the standard places (tb4.droso.net, my ports building jails)
but couldn't reproduce it.

So I'm waiting for this one to go through a run on i386 machines
(http://portsmon.freebsd.org/portoverview.py?category=archivers&portname=lzo2
shows it once happens on i386) to fail and to get a pav or erwin
email about it, which will at least show us where it goes wrong.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@219391 aed309b6-a8cd-e111-996c-001c23d10e55