Add p5-Variable-Magic, associate user-defined magic to variables from Perl.
Magic is Perl way of enhancing objects. This mechanism let the user add
extra data to any variable and overload syntaxical operations (such as
access, assignation or destruction) that can be applied to it. With this
module, you can add your own magic to any variable without the pain of the C
API.

WWW: http://search.cpan.org/dist/Variable-Magic/


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@219459 aed309b6-a8cd-e111-996c-001c23d10e55