Remove the port for Pine, and the slave port for Pico. UW has discontinued development on the Pine branch, and is instead focusing its efforts on the successor, Alpine.    http://www.washington.edu/alpine/
Alpine has more features, much better localization, better organized
(and largely rewritten) code, is BSD licensed, and should be a drop-in
replacement in terms of configuration files.

Alpine version 1.00 was released on March 17, 2008 after over a year of
development and community testing, and has since had 2 releases for
bugfixes and new features, so it's safe to say that this is a mature
replacement.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@219613 aed309b6-a8cd-e111-996c-001c23d10e55