Upgrade from 1.2.2b to 1.2.7.
Install header-files, which make the port's shared library usable by
other applications -- such as, most importantly, OpenOffice.

While here, ensure, the already installed -lintl is used, and configure
in the use of readline too.

Rephrase COMMENT to ease the confusion, that the port is only of
interest to Hungarian-speakers. It is not -- the software is an improved
version of myspell and can work with any dictionary.

Approved by:	maintainer


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@219789 aed309b6-a8cd-e111-996c-001c23d10e55