Add math/octave-forge-base.
This baseport provides the basic directory structure.  Deleting this package
will completely erase all the structure, and remaining files, associated with
the octave packaging system.  This is useful if things go wrong with the
octave packaging system.

(The octave-forge package is the result of The GNU Octave Repositry project,
which is intended to be a central location for custom scripts, functions and
extensions for GNU Octave. contains the source for all the functions plus
build and install scripts.)

PR:             127030
Submitted by:   Stephen Montgomery-Smith <stephen@math.missouri.edu>


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@220196 aed309b6-a8cd-e111-996c-001c23d10e55