Add port math/octave-forge-octgpr
This is octgpr.

The package allows interpolating and smoothing scattered
multidimensional data using Gaussian Process Regression (also known
as Kriging).

The octave-forge package is the result of The GNU Octave Repositry project,
which is intended to be a central location for custom scripts, functions and
extensions for GNU Octave. contains the source for all the functions plus
build and install scripts.

PR:             127030
Submitted by:   Stephen Montgomery-Smith <stephen@math.missouri.edu>


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@220237 aed309b6-a8cd-e111-996c-001c23d10e55