Add port math/octave-forge-communications
This is communications.

 Digital Communications, Error Correcting Codes (Channel Code),
Source Code functions, Modulation and Galois Fields

The octave-forge package is the result of The GNU Octave Repositry project,
which is intended to be a central location for custom scripts, functions and
extensions for GNU Octave. contains the source for all the functions plus
build and install scripts.

PR:             127030
Submitted by:   Stephen Montgomery-Smith <stephen@math.missouri.edu>


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@220277 aed309b6-a8cd-e111-996c-001c23d10e55