Because the many-to-many relationships are not real relationships, they can not be introspected with DBIx::Class. Many-to-many relationships are actually just a collection of convenience methods installed to bridge two relationships. This DBIx::Class component can be used to store all relevant information about these non-relationships so they can later be introspected and examined.
This module is fairly esoteric and, unless you are dynamically
creating something out of of a DBIC Schema, is probably the wrong
solution for whatever it is you are trying to do. Please be advised
that compatibility is not guaranteed for DBIx::Class 0.09000+. We
will try to mantain all compatibility, but internal changes might
make it impossible.

WWW: http://search.cpan.org/dist/DBIx-Class-IntrospectableM2M/


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@223548 aed309b6-a8cd-e111-996c-001c23d10e55