The kldpatch utility can print or alter the content of device/quirk tables in kernel modules. These tables are generally used to identify devices, and possibly apply specific quirks to enable/disable certain features.
Kldpatch is especially useful to let the kernel recognise a new device
without rebooting and rebuilding/reinstalling kernel or modules.

WWW: http://info.iet.unipi.it/~luigi/FreeBSD/

Note - this may be worth importing in the base system, however we still
need it as a port for older OS releases.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@225891 aed309b6-a8cd-e111-996c-001c23d10e55