A collection of classes for implementing web services in Objective-C
A collection of classes for making and/or parsing XMLRPC and SOAP requests.
This is intended to provide a framework for implementing web services,
so this release also contains initial support for parsing/editing
web service definitions (WSDL).

LICENSE: LGPL2 or later

WWW: http://www.gnustep.org/resources/downloads.php


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@230466 aed309b6-a8cd-e111-996c-001c23d10e55