New port devel/p5-Config-JFDI:
Config::JFDI is an implementation of Catalyst::Plugin::ConfigLoader
that exists outside of Catalyst.

Essentially, Config::JFDI will scan a directory for files matching a
certain name. If such a file is found which also matches an extension
that Config::Any can read, then the configuration from that file will
be loaded.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@231255 aed309b6-a8cd-e111-996c-001c23d10e55