Update to cmucl-19f.
After discussions with the CMUCL implementor's list it has been
decided that the FreeBSD port should stay as it is, which means that
the 6.3 binary gets installed on all FreeBSD versions.  That means the
port will not be smartened up to select binaries built for newer
FreeBSD versions that might be available.  Consensus on the list was
that those who care about that level of detail do not use the port
(remember that it's a binary distribution anyway) and that for those
who casually want to run some Common Lisp code it is good enough to
just run the 6.3 binary.

The next release will probably not have a FreeBSD-6.3 binary but at
that time the majority of users will probably be on FreeBSD-7.x or
higher.

Tested: install/deinstall, manual check for proper removal, portlint.

Commited during partial thaw as a leave port that doesn't affect other
ports.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@232950 aed309b6-a8cd-e111-996c-001c23d10e55