This is a Linux Fedora 10 infrastructure port for...
The Cyrus SASL (Simple Authentication and Security Layer)

SASL is the Simple Authentication and Security Layer, a method
for adding authentication support to connection-based protocols.
To use SASL, a protocol includes a command for identifying and
authenticating a user to a server and for optionally negotiating
protection of subsequent protocol interactions. If its use is
negotiated, a security layer is inserted between the protocol
and the connection.

WWW: http://cyrusimap.web.cmu.edu/

Obtained from:	Peter Jeremy <peterjeremy@optushome.com.au>


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@237059 aed309b6-a8cd-e111-996c-001c23d10e55