- Correct pkg-descr:
A free open source stylesheet for the Subversion index web page.
The XSLT and CSS is fully customizable. There is icons for folders
and many different filetypes. Install Repos Style as it is or use
is as a sample XSLT to design your own. A simple log viewer is also
included.

WWWW:	http://www.reposstyle.com/

Reported by:	danfe


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@240149 aed309b6-a8cd-e111-996c-001c23d10e55