For all: Add an OPTION (on by default) to install the appropriate symlinks for named.conf and rndc.key in /usr/local/etc and /var/named/usr/local/etc.
For bind9[456]:
Add OPTIONs (off by default) for the DLZ configure options, and their
corresponding ports knobs. [1] The basic infrastructure for this was
provided in the PR, but this version is slightly different in a few
details so responsibility for bugs is mine.

PR:		ports/122974 [1]
Submitted by:	Michael Schout <mschout@gkg.net> [1]


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@240544 aed309b6-a8cd-e111-996c-001c23d10e55