Add new port graphics/klatexformula:
KLatexFormula is a Qt4 program to easily get an image from a LaTeX
formula. Just enter a formula, click "Evaluate", then you can
drag&drop or copy the resulting image to another location (an
OpenOffice document, for example) or save it as an image (many
formats are available).

WWW:	http://klatexformula.sourceforge.net


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@244438 aed309b6-a8cd-e111-996c-001c23d10e55