Orc â The Oil Runtime Compiler
Orc is a library and set of tools for compiling and executing
very simple programs that operate on arrays of data.  The "language"
is a generic assembly language that represents many of the features
available in SIMD architectures, including saturated addition and
subtraction, and many arithmetic operations.

(OIL stands for Optimized Inner Loops)

WWW: http://code.entropywave.com/projects/orc/


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@250761 aed309b6-a8cd-e111-996c-001c23d10e55