- Fix mythtv on CURRENT after r203862
The commit r203862 indirectly broke mythtv because until
that time the missing include of <sys/types.h> caused
that mythtv used gettimeofday(2) instead of ftime(3) and
nobody noticed that building with ftime is broken in
mythtv.

So instead of repairing the deprecated ftime usage I've
removed it and use the gettimeofday fallback as it was
until now but by pure intention now.

Approved by:	beat (co-mentor)
Reported by:	pointyhat via erwin


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@254329 aed309b6-a8cd-e111-996c-001c23d10e55