Upgrade to version 0.8.
This represents a milestone release over the previous version, and
includes significant improvements to the version detection and
comparison algorithms, support for the new SourceForge mirror layout
(and an easy way to add future workarounds), theoretical support for
NetBSD's pkgsrc and OpenBSD ports, support for providing a list of
software in an XML file in place of ports, and support for SQLite.
Of course, there are also many optimisations and general bug fixes.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@254986 aed309b6-a8cd-e111-996c-001c23d10e55