After some mistakes i made on this port last days, i hop^W believe this is the last commit.
The original problem is port was broken on amd64 when you have CFLAGS= defined
on make.conf. On last commit i thought it was fixed, but no, i don't know why
but it was building without define LUA_USE_LINUX and the binary generated was
wrong.

I figured out just using GMAKE fixes everything, it respect CFLAGS= when you
have, add -fPIC for amd64 and use correct defines for building, so, lets use
gmake to build this port for now.

Pointyhats to:	garga@


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@261556 aed309b6-a8cd-e111-996c-001c23d10e55