Add some improvements via Doug:
This patch changes the order in which the conf file is read vs.
when the default variable assignments are made in order to provide
the ability to override them in the conf file. It also adds the
ability to include a conf file in the same directory as the script.

The patch also changes some of the = assignments to := where
having a null setting by mistake would be fatal.

The use case for these changes is the ability to "package" all
of the elements together in one place for use on multiple systems
that can all mount the same central location.

PR:		154518
Submitted by:	dougb
Hat:		secteam
Feature safe:	yes


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@268644 aed309b6-a8cd-e111-996c-001c23d10e55