- Please welcome Xorg 7.5.1
The X-Server has been patched to the latest 1.7.X series, drivers
and fonts have been updated to the latest versions. This update includes
some components from Xorg 7.6 with a lot of improvements, and it seems
that the performance is much better than the old version. Also this
Update fix build issues with clang and binutils 2.17.50 in current,
remove 6.X gruft and de-author pkg-descr.

I would like to thank:

	Beat Gaetzi
	Dima Panov
	Koop Mast
	Eitan Adler
	and all Testers.

PR:		ports/147646
		ports/148035
		ports/148744
		ports/150223
		ports/152889
		ports/154457


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@269720 aed309b6-a8cd-e111-996c-001c23d10e55