The GitDB project implements interfaces to allow read and write access to git repositories. In its core lies the db package, which contains all database types necessary to read a complete git repository. These are the LooseObjectDB, the PackedDB and the ReferenceDB which are combined into the GitDB to combine every aspect of the git database.
For this to work, GitDB implements pack reading, as well as loose object
reading and writing. Data is always encapsulated in streams, which allows
huge files to be handled as well as small ones, usually only chunks of
the stream are kept in memory for processing, never the whole stream at
once.

WWW: http://packages.python.org/gitdb/


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@270984 aed309b6-a8cd-e111-996c-001c23d10e55