http://www.vdr-wiki.de/wiki/index.php/Upnp-plugin
UPnP/DLNA Plugin for Video Disk Recorder

This Plugins extends the VDR with the possibility to act as an UPnP/DLNA Media
Server (DMS). It will serve VDR's contents in the network to any UPnP-AV and
DLNA capable devices.

This still is an alpha version!

WWW: http://upnp.vdr-developer.org/


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@271809 aed309b6-a8cd-e111-996c-001c23d10e55