ISC DHCP 4.2.x includes features that were not included in DHCP 4.1.x. These include:
Processing the DHCP to DNS server transactions in an asynchronous fashion.
The DHCP server or client can now continue with it's processing while
awaiting replies from the DNS server.

There are a number of DHCPv6 limitations and features missing in this
release, which will be addressed in the future:

- DHCPv6 includes human-readable text in status code messages, in
English. A method to reconfigure or support other languages would
be preferable.

- The "host-identifier" option is limited to a simple token.

- The client and server can only operate DHCPv4 or DHCPv6 at a time,
not both. To use both protocols simultaneously, two instances of the
relevant daemon are required, one with the '-6' command line option.

PR:		ports/158516
Submitted by:	Ryan Steinmetz <rpsfa@rit.edu>


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@277279 aed309b6-a8cd-e111-996c-001c23d10e55