Add openscad. OpenSCAD is a software for creating 3D CAD objects.
Unlike most free software for creating 3D models, it does not focus on
the artistic aspects of 3D modelling but instead on the CAD aspects.
Thus it is more suitable for creating 3D models of machine parts but
perhaps not when creating computer-animated movies.

OpenSCAD is not an interactive modeller. Instead it is something like
a 3D-compiler that reads in a script file that describes the object
and renders the 3D model from this script file. This gives designers
full control over the modelling process and enables them to easily
change any step in the modelling process or make designes that are
defined by configurable parameters.

WWW: http://www.openscad.org/

Obtained from (slightly):	OpenBSD


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@280009 aed309b6-a8cd-e111-996c-001c23d10e55