- New port math/octave-forge-secs3d.
The octave-forge package is the result of The GNU Octave Repositry project,
which is intended to be a central location for custom scripts, functions and
extensions for GNU Octave. contains the source for all the functions plus
build and install scripts.

This is secs3d.  A Drift-Diffusion simulator for 3d semiconductor devices.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@281674 aed309b6-a8cd-e111-996c-001c23d10e55