Fix build when boost-libs is compiled with a newer (> 4.2 from base) gcc.
LyX originally uses the AC_CHECK_LIB macro to check for boost, however
this macro always calls a C compiler, even to check for a C++ library
such as boost. This made the test program built for this check to
always fail to link when boost was compiled witha recent enough
compiler.

A fix was submitted upstream [1], and this patch needs to be applied
meanwhile.

[1] http://article.gmane.org/gmane.editors.lyx.devel/138405

Reported by:	avg
Approved by:	avilla (mentor)


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@281873 aed309b6-a8cd-e111-996c-001c23d10e55