- Add p5-Excel-Writer-XLSX 0.26
The Excel::Writer::XLSX module can be used to create an Excel file in the 2007+
XLSX format. The XLSX format is the Office Open XML (OOXML) format used by Excel
2007 and later. Multiple worksheets can be added to a workbook and formatting
can be applied to cells. Text, numbers, and formulas can be written to the
cells. This module cannot, as yet, be used to write to an existing Excel XLSX
file.

WWW: http://search.cpan.org/dist/Excel-Writer-XLSX/


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@282816 aed309b6-a8cd-e111-996c-001c23d10e55