Grant graphics/linux-tiff a stay of execution
Due to the way that linux dependencies are recorded my greps for the
portname, category/portname, etc. didn't discover that on RELENG_7
this version of the port is still heavily depended on. Neither did
my building INDEX on more recent versions of FreeBSD.

So bring it back for now, with the plan to mark those ports that
depend on it BROKEN conditional on OSVERSION.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@283549 aed309b6-a8cd-e111-996c-001c23d10e55