- Fix definition of PREFIX and NO_MTREE.
As noted in the file, PREFIX and NO_MTREE need to be defined in the
pre-makefile section, as PREFIX is checked prior to the end of the
section (in bsd.port.mk), and NO_MTREE in the post-makefile section
(in bsd.port.mk and other *.mk files).

The "error" (it was actually an annoyance, but could have been worse)
was triggered when updating KDE ports using the new WITH_PKGNG switch
(bsd.pkgng.mk, also bugged in a similar way at the time, resolved the
value of NO_MTREE prior to its definition in bsd.kde4.mk).


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@292175 aed309b6-a8cd-e111-996c-001c23d10e55