New port: (devel/py-importlib)
This package contains the code from importlib as found in Python 2.7.

It is provided so that people who wish to use importlib.import_module()
with a version of Python prior to 2.7 or in 3.0 have the function readily
available. The code in no way deviates from what can be found in the 2.7
trunk.

For documentation, see the importlib docs for Python 2.7.

WWW: http://pypi.python.org/pypi/importlib/1.0.

PR:	ports/165420
Submitted by:	Alexandros Kosiaris <akosiaris@gmail.com>


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@292190 aed309b6-a8cd-e111-996c-001c23d10e55