Add a proper description.
The main function of wiggle is to apply a patch to a file in a similar
manner to the patch(1) program.

The distinctive difference of wiggle is that it will attempt to apply a
patch even if the before part of the patch doesn't match the target
file perfectly. [...]

Feature safe: yes


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@293959 aed309b6-a8cd-e111-996c-001c23d10e55