The samba rc.d script uses some clever tricks to start (up to) 3 different services using the same script. As a result it resets rcvar several times in order to process the options for each service.
The changes I made on 2012/01/14 to facilitate the removal of set_rc_var()
from HEAD were effective in the case where the WINBIND option was off (the
case that I tested) because that causes the related portions of the rc.d
script to be removed completely on install. However, if installed from a
package, or installed using the the default OPTIONS, WINBIND is on, which
caused the last known rcvar to be winbind_enable.

Since the common case seems to be for users to use samba_enable (which
only enables smb_and nmb_ by default) the fact that rcvar=winbind_enable,
but that knob is off, caused the startup script to trip on a totally
unrelated portion of rc.subr.

So the fix is to move processing of the winbind_ stuff first, which leaves
the last known rcvar as smb_enable. Since running nmb without smb is a
very unlikely scenario, this should be safe for the common case, as well
as safe if the user enables winbind_.

Apologies all around for not catching this sooner, and thanks to the users
who reported the problem and stuck with me while I debugged it.

Bump PORTREVISION since this fix is needed for the common case, as
configured for the package.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@294765 aed309b6-a8cd-e111-996c-001c23d10e55