- Add p5-Compress-Snappy 0.18
Compress::Snappy provides an interface to Google's Snappy (de)compressor.

Snappy does not aim for maximum compression, or compatibility with any other
compression library; instead, it aims for very high speeds and reasonable
compression. For instance, compared to the fastest mode of zlib, Snappy is an
order of magnitude faster for most inputs, but the resulting compressed files
are anywhere from 20% to 100% bigger.

WWW: http://search.cpan.org/dist/Compress-Snappy/


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@295048 aed309b6-a8cd-e111-996c-001c23d10e55