The KDE/FreeBSD team is pleased to announce version 2.5 of Calligra, coming to the ports tree after a short testing period which showed it to be quite stable.
For the productivity part of the suite (word processor, spreadsheet,
and presentation program) the target user of version 2.5 is still the
student or academic user. This version has a number of new features
that will make it more suitable for these users.
The artistic applications of the Calligra Suite are the most mature
ones and are already used by professional users everywhere.

As usual, detailed release notes can be found in the official
announcement:
http://www.calligra.org/news/calligra-2-5-released

Meanwhile, new translations were added:
- editors/calligra-l10n-gl (Gallegan)


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@303149 aed309b6-a8cd-e111-996c-001c23d10e55