Update to the 20120922 snapshot of GCC 4.7.3, which is pretty close to the GCC 4.7.2 release.  It also fixes an ABI problem introduced with earlier GCC 4.7 releases.
Cut down the comments at the top of Makefile per recent guidance.

Revamp the handling of different languages (frontends, run-times)
by GCC to allow for orthogonal setting them in the future. [1]

Print the list of languages being built as part of pre-everything
and shorten the output of operating system and version there.

Inspired by:	jkim [1]


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@304996 aed309b6-a8cd-e111-996c-001c23d10e55