Work around a problem on stable/[89] with libotr generating an "undefined reference to `__stack_chk_fail_local'" error. None of the usual remedies work (such as making sure that gcc is used instead of ld for the linker) so on those releases we simply disable that option.
pointyhat logs confirm that pidgin-otr (the only consumer of libotr
atm) is failing on 8 and 9 with the same configure error that I am
seeing on 8, so this patch should at least allow it to build on those
releases.

Bump PORTREVISION for libotr to err on the side of caution.

While I'm here, remove a now-spurious mod to the pidgin-otr configure.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@305438 aed309b6-a8cd-e111-996c-001c23d10e55