- New port math/octave-forge-database.
The octave-forge package is the result of The GNU Octave Repositry project,
which is intended to be a central location for custom scripts, functions and
extensions for GNU Octave. contains the source for all the functions plus
build and install scripts.

This is database: interface to SQL databases, currently only postgresql using
libpq.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@311535 aed309b6-a8cd-e111-996c-001c23d10e55