Fix installation of devel/eric4 when stdin is closed.
The installation asks the user a question, but if stdin is closed an EOFError is
thrown and installation fails.  The patch fixes this by assuming "no" to the
question.

The install script asks the user about modifying PyXML, an external file to
eric4, thus answering "no" is the appropriate answer in the absence of user
judgement.

Approved by:	eadler (mentor), bsam (maintainer)


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@312630 aed309b6-a8cd-e111-996c-001c23d10e55