add new port: devel/py-billiard
billiard is a fork of the Python 2.7 multiprocessing package.
The multiprocessing package itself is a renamed and updated version
of R Oudkerk's pyprocessing package. This standalone variant is intended
to be compatible with Python 2.4 and 2.5, and will draw it's
fixes/improvements from python-trunk.

WWW: https://github.com/celery/billiard

PR:		176663
Submitted by:	william88@gmail.com


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@314025 aed309b6-a8cd-e111-996c-001c23d10e55