add new port: games/f1lt
F1LT is a free unofficial Formula 1 live timing application written for
Windows, Linux, Symbian and Android systems. It is an alternative
application to the java applet available on the formula1.com site.
To use the application you only need an account on the formula1.com
site, the same you are using for connecting to Live Timing from formula1.com.

Feature list is too long to list here, but please check WWW address for more!

WWW: http://f1lt.pl/

Just in time for Malaysia!


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@315082 aed309b6-a8cd-e111-996c-001c23d10e55