Add creduce:
C-Reduce is a tool which takes a large C or C++ program that has a property of
interest (such as triggering a compiler bug) and automatically produces a much
smaller C/C++ program that has the same property. It is intended for use by
people who discover and report bugs in compilers and other tools that process
C/C++ code.

Reviewed by:	mva, swills
Tested by:	swills


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@316494 35697150-7ecd-e111-bb59-0022644237b5