Add Gstreamer 1.0.7. This is the new stable version of the Gstreamer multimedia framework.
Use the USE_GSTREAMER1 macro to select Gstreamer 1.0 plugins. A new chapter
for the porters handbook about gstreamer is a work in progress.

X related plugins are now stand alone. The gio plugin is now part of the
gstreamer1-plugins port and the oss plugins are now part of the
gstreamer1-plugins-good port.

Note that the pulseaudio plugin is missing because it needs a newer version
of pulseaudio. Also there are still a number of plugins from -bad which
haven't been ported to the 1.0 API yet.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@316693 35697150-7ecd-e111-bb59-0022644237b5