Set explicit comment for slave port emulators/i386-wine-devel.
i386-wine-devel is the normal wine-devel port built and packaged such that it
can be installed (via `pkg add` or `pkg_add`) in an amd64 environment.  It can
only be built in an i386 environment (hence the retained ONLY_FOR_ARCH) as
the Ports Collection does not support cross compiling.

Approved by:	eadler (mentor)


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@317425 35697150-7ecd-e111-bb59-0022644237b5