Update to GCC 4.6.4, the final release and end of the GCC 4.6 branch.
(This requires an EPOCH bump due to how snapshots are labeled.)

Introduce a new option BOOTSTRAP that allows for doing a full
bootstrap of GCC, as opposed to just running a simple build.
Bootstrapping is actually the default upstream, we disable it
by default for the stable flavors of GCC since that is a huge
win in terms of build time of the port.  No change in default
behavior for this port.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@317986 35697150-7ecd-e111-bb59-0022644237b5