Add the patch I added to qt4-corelib in r317896 to some other Qt ports.
These ports which previously failed to build with clang-i386 use their
private copies of the qsimd_p.h header, since this private header is not
installed at all (which, in turns, makes the PORTREVISION bump in
qt4-corelib quite useless).

We thus need to add a copy of the same patch to all these ports to make sure
the qsim_p.h they use when being built is actually the one we want.

Hopefully this gets miwi happy and people on clang-i386 can build these
ports again.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@318239 35697150-7ecd-e111-bb59-0022644237b5