add new port: www/cakephp23
This is a port of the CakePHP web application framework. CakePHP is a
free open-source rapid development framework for PHP. It's a structure
of libraries, classes and run-time infrastructure for programmers
creating web applications originally inspired by the Ruby on Rails
framework. Our primary goal is to enable you to work in a structured
and rapid manner - without loss of flexibility.

Add conflicts to the various cakephp ports.

PR:		ports/176130
Submitted by:   Phil Phillips <pphillips@experts-exchange.com>
		(maintainer of other cakephp ports)


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@321716 35697150-7ecd-e111-bb59-0022644237b5