Add some notes about rebuilding kdelibs4-dependent ports if using clang.
As mentioned in r322300, clang support has been improved since the update to
4.10.4 a few days ago.

This, however, requires a rebuild of the ports that depend on kdelibs4. Most
of them are covered by the KDE update itself, but those which are not part
of the Software Compilation need to be rebuilt manually to make sure the
previous issues (proper symbol visibility being the most annoying of them)
are solved.

I haven't fully tested all the possible portmaster/portupgrade/pkg commands
myself, so please shout out loud if they're not correct.

In collaboration with:	avilla@


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@322343 35697150-7ecd-e111-bb59-0022644237b5