www/phpmustache: Unbreak fetch and update to version 2.4.0
This port had some sort of broken roll-your-own github fetch support.  It
could have been easily corrected for the current version 0.8.1, but that
was 11 releases ago!

So in addition to fixing the fetch, the port was updated to the latest
release v2.4.0.  The port now requires PHP to build.  It always required
PHP as a run depends but this was previously omitted.  The MIT license was
moved to the built-in license handling, and the documentation was eliminated
because it is in unreadable markdown format.  The test files were also
eliminated because it doesn't appear to be meant for users.  The examples
were kept and augmented.  The name of the main class file changed from
"Mustache.php" to "mustache.php" but gets installed to the same location.

PR:		ports/179961
Approved by:	bapt (mentor), maintainer timeout (~4 weeks)


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@323492 35697150-7ecd-e111-bb59-0022644237b5