devel/afay: Move to Ada framework
Currently this port has a hard dependency on lang/gnat.  This has been
replaced with USES+=ada and minor changes to incorporate it into the Ada
framework with a PORTREVISION bump.

The makefile header was trimmed and NOPORTDOCS was replaced with the DOCS
option, and now portlint is happy.

Approved by:	bapt & culot (mentors)


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@323753 35697150-7ecd-e111-bb59-0022644237b5