Replace USE_GCC=4.6+ and USE_GCC=4.4+ by USE_GCC=yes.
Right now this is a noop in the former case and a noop in the latter
case unless lang/gcc44 has been installed explicitly.

This puts a bit more emphasis on standardizing on a canonical version
"current" GCC and makes it easier to update that canonical version
by changing the default in Mk/bsd.gcc.mk and updating the lang/gcc port.

That is, USE_GCC=yes means "use a decent/modern version of GCC" without
having to worry about details.

Approved by:	portmgr (bdrewery)


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@324851 35697150-7ecd-e111-bb59-0022644237b5