audio/dumb: adjust for new devel/allegro
A month elapsed between the generation of the previous commit and when it was
finally committed.  During that time, devel/allegro had a major update which
invalidated part of the previous commit, causing breakage of audio/dumb-allegro.
This commit corrects the previous one (carrying over the same approval.)

Approved by:	culot (mentor, implicit)


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@324869 35697150-7ecd-e111-bb59-0022644237b5