Fix build.
Per ISO C++03 14.2/4:

When the name of a member template specialization appears after . or ->
in a postfix-expression, or after nested-name-specifier in a qualified-id,
and the postfix-expression or qualified-id explicitly depends on a
template-parameter (14.6.2), the member template name must be prefixed by
the keyword template. Otherwise the name is assumed to name a non-template.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@326456 35697150-7ecd-e111-bb59-0022644237b5