www/metacafe_dl: Unbreak fetch (sort of)
The previous commit removed the blanking of EXTRACT_SUFX and EXTRACT_ONLY
thinking that it was a mistake.  The distfile actually has no extension,
so this is really needed.  Without it, the checksum can't even start.
The MASTER_SITES is gone.  The file is still fetchable because it is cached
but at some point it is going to disappear unless somebody preserves it.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@327171 35697150-7ecd-e111-bb59-0022644237b5