lang/php5*: Fix spl/simplexml modules (4 ports)
The simplexml module uses spl which uses pcre (php53,54,55).  For php52,
the fix is applied directly to spl.  In any case, the bundled PCRE is not
used, so it needs to be able to find the external pcre library and related
headers.  Unfortunately, there is no include for ${LOCALBASE}/include
directory.  This commit adds to fix simplexml (or spl) module build.

Approved by:	portmgr (bapt)


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@327187 35697150-7ecd-e111-bb59-0022644237b5