Add sysutils/nfsping
NFSping is a command line utility for measuring the response time of an NFS
server. It's basically a copy of the fping interface but doesn't share any code
with that project.

On modern NFS servers, the network stack and filesystem are often being run on
separate cores or even hardware components. This means in practise that a fast
ICMP ping response isn't indicative of how quickly the NFS filesystem is
responding. This tool more directly tests the responsiveness of the server's
operating system's NFS component.

WWW: https://github.com/mprovost/NFSping


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@330948 35697150-7ecd-e111-bb59-0022644237b5