multimedia/mplayer2: fix configure failure, indirect linking
As seen when using gcc48 as the ports compiler (and on DragonFly),
mplayer2 was misconfigured.  The avresample library doesn't exist
during configuration, thus it will fail when using a modern
linker that complains that the library isn't found.

Moving -lavresample from extra-ldflags to extra-libs fixes the
problem on both FreeBSD+gcc48 and DragonFly

Approved by:	portmgr (bapt, implicit)


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@332095 35697150-7ecd-e111-bb59-0022644237b5