Add firstboot-pkgs:
When the system first boots, install the pkg(8) tools (if not already
installed) and packages listed in the $firstboot_pkgs_list rc.conf
variable.

Obviously, this port is not useful after a system is already running; it is
intended to be included as part of the installation or disk image building
process.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@335251 35697150-7ecd-e111-bb59-0022644237b5