Fix build of dvbcut , which was broken by a recent ffmpeg upgrade.
This port is no longer based on the Qt3 version of dvbcut plus a large
patch to make it build on Qt4 with the Qt3 emulation layer. A fork that
uses native Qt4 calls is used instead.

The generated files moc_progressstatusbar.cpp and moc_progresswindow.cpp
trigger an internal error in clang, if compiled with optimization. This
bug has been accepted as 18098 in the llvm bug tracker. To work around
this problem, the above mentioned two files are compiled with -O0 appended
to the compiler flages. The post-configure target in the ports Makefile is
to be removed, when a fixed version of clang is distributed in all FreeBSD
releases supported by the ports infrastructure.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@335469 35697150-7ecd-e111-bb59-0022644237b5