Add sanity check for LIB_DEPENDS and RUN_DEPENDS.
Ports does not support relative paths in ${TYPE}_DEPENDS and actually breaks
the dependency registration with pkgng.  This occurs in ACTUAL_PACKAGE_DEPENDS
where ${PORTSDIR} is stripped from the dependencies' directory and uses the
rest of the string as the package origin.  pkg(8) then fails to detect package
origins with relative paths.

Approved by:	portmgr (bapt)


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@335512 35697150-7ecd-e111-bb59-0022644237b5