www/py-django-dpaste: Add option for MIGRATIONS support (via south)
A few migrations come packaged with this Django application and the author
notes they are "preferred", but not required.

Regarding the option naming, it was furiously debated in #freebsd-python and
I opted for naming the "feature that is enabled" rather than the application
enabling it. Another data point is the common use of the term "schema
migrations" in Django parlance.

Thank you xmj for being a great sounding board as always!

Reported by:	Johannes Meixner (xmj, via IRC)
Reviewed by:	Johannes Meixner


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@336056 35697150-7ecd-e111-bb59-0022644237b5