Update to 2.3.1
   * Staging support
   * New style LIB_DEPENDS
   * Use OPTIONS helpers
   * OPTIONS_SINGLE to select between Cyrus SASL and gsasl
   * Two new options: EXPERIMENTAL and SUPERCEDED

The port requires OpenSSL >= 1.0.1, so set WANT_OPENSSL_PORTS=yes
everywhere except for on FreeBSD 10.x, where it's in the base system.

The github.org MASTER_SITE a) redirects to a server in the Amazon
cloud and b) requires use of HTTPS.

(a) implies setting fetch(1) to follow redirects, which is usually
frowned upon.  However, this pattern is recommended in bsd.sites.mk
for use with GitHub.

(b) is troublesome for FreeBSD 8.x, as fetch(1) hasn't grown HTTPS
support there yet.  Hence provide a copy of the distfile on
http://people.freebsd.org/ as a 2nd MASTER_SITE

Approved by:	mm (maintainer)


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@336590 35697150-7ecd-e111-bb59-0022644237b5