science/elmer(post,-matc): Use modern fortran, unbreak, undeprecate
These ports are due to be deleted in a couple of days because they use
gcc34.  I was curious if they actually specifically needed gcc34 or if
any recent gfortan would do.  The answer is these ports build fine with
USES+=fortran, which pulls in lang/gcc rather than the deprecated lang/gcc34.

The elmerpost port was broken on amd64; this is because it needs the -fPIC
flag.  I built it successfully in poudriere on FreeBSD 9.2, another platform
that supposed elmerpost can't build on.  I did not test i386, we'll see what
QAT says.

The listed maintainer has been unresponsive for months on many ports, so due
to the fact these two ports are scheduled for deletion on Dec 27, I am not
getting prior approval from maintainer.  The deprecation and expiration
settings are removed.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@337470 35697150-7ecd-e111-bb59-0022644237b5