Add x11-fonts/bitter.
A "contemporary" slab serif typeface for text, especially designed for
comfortably reading on any computer or device. The robust design started from
the austerity of the pixel grid, based on rational rather than emotional
principles. It combines the large x-heights and legibility of the humanistic
tradition with subtle characteristics in the characters that inject a certain
rhythm to flowing texts.

WWW: http://www.huertatipografica.com/fonts/bitter-ht


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@337780 35697150-7ecd-e111-bb59-0022644237b5