Add DropBox & BOX remote file storage support and enable on by default
Remove NO_STAGE
Remove MAKE_JOBS_UNSAFE

Remove BUILTINTAGLIB, this is only relevant for distributions that
have a taglib package < 1.8

Remove SPOTIFY support.  Requires vendor binary library that does
not currently support FreeBSD.  The QCA code is a tool to fetch this
library if enabled.

Remove -DIMOBILEDEVICE as its not a Cmake build options for this port

Reviewed by:	vg


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@338393 35697150-7ecd-e111-bb59-0022644237b5