french/alphabet_sounds_fr: Fix and support stage
The PR was written because the port started to fail in jail where $HOME
is read-only.  It turns out that it is not the port's fault, but rather
that of the (former) build dependency, childsplay.

The only reason it was a build dependency was to check if it's version
is at least 0.71.  Childsplay is currently at version 2.5, so there's
no reason to check the version.  The fix was to remove the check and
childsplay as a build dependency.

Stage support was also added, so there's not much left of the original
install.sh.

PR:		ports/185202
Approved by:	maintainer timeout


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@339815 35697150-7ecd-e111-bb59-0022644237b5