- Add new port devel/llnextgen
LLnextgen is an Extended-LL(1) parser generator. It is a rewrite of the LLgen
parser generator by D. Grune and C.J.H. Jacobs which is part of the Amsterdam
Compiler Kit (ACK). Like all parser generators, LLnextgen takes the description
of the grammar with associated actions as input, and generates a parser routine
for use in compilers and other text processing programs.

PR:		ports/185458
Submitted by:	Oleksii Tsai <oleksii.tsai@googlemail.com>


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@340087 35697150-7ecd-e111-bb59-0022644237b5