Partially fix the build.
Sphinx underwent an incompatible change in version 1.2 that removed the
sphinx.ext.refcounting module. The module is referenced by shiboken's Sphinx
configuration file, and the build fails.

Fix it by removing the reference to sphinx.ext.refcounting, as it does not
seem to be required and was probably added by default when the configuration
file was generated.

As mentioned, this is a partial fix: platforms that use libc++ by default
are still broken due to other issues, but at least FreeBSD 8 and 9 should
have a working package again.

While here, also use the new syntax for LIB_DEPENDS.

PR:		ports/185179
Submitted by:	rakuco@
Approved by:	maintainer timeout (27 days)
MFH:		2014Q1


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@340659 35697150-7ecd-e111-bb59-0022644237b5