Introduce USES=libtool to replace USE_AUTOTOOLS=libtool.
Unlike USE_AUTOTOOLS=libtool which makes a port use libtool from
devel/libtool, this just patches the copy of libtool included in the
port.  This allows adding it to all ports that use libtool including
ports that build Fortran libraries and ports that use different
compilers or binutils than the ones used to build devel/libtool.

USES=libtool also changes the library naming specification similar
to what USE_GNOME=ltverhack does.  Given the libtool version info
$current:$revision:$age and $major=$current-$age the library will be
named libname.so.$major.$age.$revision instead of libname.so.$current
and libname.so.$major becomes a symlink.  Because $major increases
less frequently than $current this reduces the number of library
version bumps which means fewer rebuilds of dependent packages.

To ease the conversion from USE_AUTOTOOLS=libtool this naming spec
can be disabled with USES=libtool:oldver but the intention is that
all libtool ports eventually use the new naming.

Another change is that only libraries that are listed on the command
line will be linked into executables.  Normally libtool would link in
the entire tree of library dependencies.  This reduces the number of
direct dependencies on a given library which means fewer packages
need to be rebuilt on a library version bump.

Approved by:	portmgr (bapt)


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@341646 35697150-7ecd-e111-bb59-0022644237b5