devel/spark2012: Build the static library to stop ironsides leftovers
Without a static library present, gnatmake will build the objects it
needs in the base system -- that's not legal.

The original issue I had with this was using BSD_INSTALL_LIB which
stripped out the symbols in the static library, but I didn't realize
what was happening.  Now I know and the library is installed intact
with BSD_INSTALL_DATA.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@344986 35697150-7ecd-e111-bb59-0022644237b5