net-mgmt/rancid, devel/cvs(2html,-syncmail): Fix recent cvs dependency
CVS is no longer in base on FreeBSD 10 and later, but it's still in
DragonFly's base, and the devel/cvs port doesn't even build on DF.  The
dependence on CVS was added to some ports recently, make sure that only
applies to FreeBSD.  Note the rancid approach where it further limits
it to FreeBSD 10+ is better than devel/cvs* where they pull in CVS
unconditionally on FreeBSD 8 and 9 where it's not needed.  The
unmaintained cvs-syncmail port was updated to be like rancid, but I
didn't make the same update to the maintained cvs2html port.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@346745 35697150-7ecd-e111-bb59-0022644237b5