lang/asis: New version as result of gcc-aux upgrade to gcc49
The ASIS source has to be matched with the compiler source of the compiler
that builds it.  ASIS is always tricky because it's matched to annual
releases of GNAT GPL, and never to GNAT FSF.  However, after disabling the
useless (and faulty) strong version check, ASIS 2013 seems to work okay
with gcc49.  At least the two ports that use it seem to work ok.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@348174 35697150-7ecd-e111-bb59-0022644237b5