lang/gcc-aux: Fix parallel jobs build (GCC bug)
The cause for the build errors are due to a GCC makefile bug.
C++ is now required to build gnattools, but libstdc++ was not listed
as a dependency.  The fix for this bug will be posted to GCC's bugzilla
shortly.  While here, bump gcc-aux to the latest snapshot (16 March).
Now gcc-aux passes Redports 8x.

Reported by: pkg-fallout


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@348790 35697150-7ecd-e111-bb59-0022644237b5