APQ ports: apq, apq-mysql, apq-pgsql, apq-odbc
Without direction, GPRconfig will create a cgpr file at the standard
$PREFIX/lib/gnat repository.  This is the cause of the builder
failures because it touches a part of the userland it's not supposed to.

GPRconfig should be modified not to do this; it is never desirable.
In the meantime, the fix for APQ is to specify .cgpr targets which do
not exist so GPRconfig creates them there instead of $PREFIX/lib/gnat.

Reported by: swills


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@349205 35697150-7ecd-e111-bb59-0022644237b5