ocaml update 3/4: devel/omake major tweaks, stage support
Two dozen patches were brought in from NetBSD via dports , and this port
was staged too. Most of the patches are upstream fixes as Omake has been
updated in repo but without a new release.  In any case an update was
required because it broke after Ocaml was updated to version 4.00.

Urged by: portmgr (bapt)


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@349274 35697150-7ecd-e111-bb59-0022644237b5