www/aws: Fix usage of aws in jailed environment
AWS has a sanity check that assumes that binding to the standard loopback
address of 127.0.0.1 is always possible, but this is not a good assumption
inside a FreeBSD jail.  The result is that connection is success because
it adjusts the 127.0.0.1 address correctly on the fly, but the sanity
check is no longer valid.  The provided patch changes the sanity check to
get the address rather than assuming 127.0.0.1.

PR:		ports/188342
submitted by:	Natacha Porte
Approved by:	maintainer (myself)


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@351204 35697150-7ecd-e111-bb59-0022644237b5