Add USE_GNOME+= gnomedocutils to 5 ports
These ports will fail configuration if gnome-doc-utils is not present.
Likely a common dependency loaded it in the past, but not anymore.
Explicitly list gnome-doc-utils as a build dependency in order to
fix the recent breakage of these ports.

Also reset maintainership of etherape - maintainer gone for more than
one yet now.

Approved by:	portmgr (bdrewery)


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@351410 35697150-7ecd-e111-bb59-0022644237b5