lang/ocaml: Fix NO_PROFILE setting and native compiler for DragonFly
The ocaml port was not respecting a NO_PROFILE definition in make.conf,
which causes a stage check failure (missing files) if it is set. Improve
that logic and patch the configure script in numerous places to support
building the native code compiler on DragonFly.  This is one of those
famous ports that think all BSDs have the pattern "bsd" in their target
triplet.

Continued under PR: ports/188158


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@353142 35697150-7ecd-e111-bb59-0022644237b5