www/piwik: Upgrade version from 2.2.1 => 2.2.2 to unbreak
This port has a history of rerolling and upstream has stated to the
maintainer they don't promise to stop this practice, only minimize it.
I made an appeal to them to cease the practice, but if they refuse this
port needs to be handled differently.

Option 1: Change the MASTER_SITE to lock in the release
Option 2: Remove the port as constantly changing tarballs is unsuitable

For the moment, piwik builds again.  The tarball changed after the PR
was initially submitted!

PR:		ports/189457
submitted by:	maintainer (Hans Nordhaug)


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@353786 35697150-7ecd-e111-bb59-0022644237b5