Fix breakage introduced by previous commit.
I have compared the two distfiles (old and new) and the only difference
was a rename of the containing directory and removal of a few test and
scratch files.

While I'm there also make pkg-message show what the administrator should
do in a new installation because the scripts are intentionally installed
under root:wheel.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@353885 35697150-7ecd-e111-bb59-0022644237b5