lang/gcc-aux: Bring in latest diffs-*
I should have done this when I bumped the port port earlier.  A few
of the changes to the diff-* files directly affect FreeBSD although
most are results of other platforms (NetBSD, OpenBSD mainly)

The most invasive change was exchanging strcpy and printf for their
"n" versions.  It was to make OpenBSD happy but the code is better
for it.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@353999 35697150-7ecd-e111-bb59-0022644237b5