RapidXml is an attempt to create the fastest XML parser possible, while retaining useability, portability and reasonable W3C compatibility. It is an in-situ parser written in modern C++, with parsing speed approaching that of strlen function executed on the same data.
RapidXml has been around since 2006, and is being used by lots of
people. HTC uses it in some of its mobile phones.

If you are looking for a stable and fast parser, look no further.
Integration with your project will be trivial, because entire library
is contained in a single header file, and requires no building or
configuration.

WWW: http://rapidxml.sourceforge.net/


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@354008 35697150-7ecd-e111-bb59-0022644237b5