- New port: devel/catch
Catch stands for C++ Automated Test Cases in Headers and is a multi-paradigm
automated test framework for C++ and Objective-C (and, maybe, C). It is
implemented entirely in a set of header files, but is packaged up as a single
header for extra convenience.

WWW: https://github.com/philsquared/Catch


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@355359 35697150-7ecd-e111-bb59-0022644237b5