math/spooles*: specify post-extraction with ${TAR}
It's possible that EXTRACT_CMD won't be predefined in the near future
in order to support distfiles in multiple formats.  We know the post-
extraction tool needs to be tar, so let's specify it directly.

It may be worth looking at the post-extract target here and seeing
if it's really needed.  Often one can just change the WRKSRC value to
avoid a custom post-extract taraget.

Approved by:	infrastructure improvements blanket


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@355754 35697150-7ecd-e111-bb59-0022644237b5