- Add p5-PkgConfig 0.07720
PkgConfig provides a pure-perl, core-only replacement for the pkg-config
utility.

This is not a description of the uses of pkg-config but rather a description of
the differences between the C version and the Perl one.

While pkg-config is a compiled binary linked with glib, the pure-perl version
has no such requirement, and will run wherever Perl ( >= 5.6 ) does.

The main supported options are the common --libs, --cflags, --static, --exists
and --modversion.

WWW: http://search.cpan.org/dist/PkgConfig/


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@357320 35697150-7ecd-e111-bb59-0022644237b5