Depend on devel/xdg-utils instead of devel/py-qt4-assistant.
This version of Eficas has dropped the dependency on py-qt4-assistant, which
has been deprecated in PyQt4 since 2010, and replaced it with a call to
xdg-open.

Follow suit and replace the runtime dependencies, paving the way for us to
get rid of py-qt4-assistant ourselves.

CR		D226
Approved by:	thierry (maintainer)


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@357906 35697150-7ecd-e111-bb59-0022644237b5