java/eclipse: Remove troublesome comment completely
Freshports won't update the entry due to choking on the MAVEN_OPTS
comment.  Truthfully the comment is more trouble than it's worth.  First,
it's MVN_OPTS, not MAVEN_OPTS.  Secondly, "+=" is a no-op, so the
definition should be there anyway.  Thirdly, MVN_OPTS is undefined.
Just remove the whole mess.

PR: 188659


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@359325 35697150-7ecd-e111-bb59-0022644237b5