net/wowzamediaserverpro: Reset maintainer and deprecate interactive port
Unstaged ports are being reset today.  I'll grab this one from the batch
in order to set the expiration date to match the other unmaintained
interactive ports.  The port should be converted to a regular port as
well as providing stage support in order to undeprecate it.

Approved by:	portmgr (implicit)


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@360321 35697150-7ecd-e111-bb59-0022644237b5