games/atomix: never-ending stage, take 3
Newer poudriere (read: current FreeBSD version) has an additional plist
check that atomix failed.  It turns out that the original port moved the
atomix.scores creation to the pkg-install script, but it never surpressed
the original Makefile code that created it.  In effect, the original
Makefile was creating the file and the pkg-install script wasn't doing
much of anything.  This caused the signature to be confusing.

The solution is simple: Surpress the vendor makefile target that creates
atomix.scores and let pkg-install create it as intended.

Diagnosis by:	bdrewery


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@361338 35697150-7ecd-e111-bb59-0022644237b5