games/moonbuggy: Fix touching $HOME during build
It's not permissible for a port to touch $HOME during the build.  The
moon-buggy game does this during install to pre-create the high scores
file $HOME/.mbscores.

Moon-buggy creates two versions of high scores: One in $HOME and the
other in /var/games/moon-buggy.  This makes compliance tricky because
the high scores at /var/games needs to persist after the package is
uninstalled if the scores have changed.

To handle this, I moved scores creation to pkg-install script and also
created a "baseline" scores files so during pkg-deinstall it can be
determined if new high scores exist.  The complimentary pkg-deinstall
script will compare the current high-score file with the baseline. If
they are the same, both are removed along with /var/games/moon-buggy
folder which makes poudriere happy.

While here:
 * Remove the unnecessary NEED_ROOT
 * Remove the unnecessary setgid configure argument
 * use @owner, @group in pkg-plist instead
 * move CHMOD to post-install with BINMODE although @mode could have
   been used as well.

This work is covered by the "Just fix it" blanklet.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@362256 35697150-7ecd-e111-bb59-0022644237b5